% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airline.R
\docType{data}
\name{airline}
\alias{airline}
\title{Airline}
\format{
A data frame with 10 rows and 6 variables:
\describe{
  \item{x1}{input one}
  \item{x2}{input two}
  \item{y1}{output one}
  \item{y2}{output two}
  \item{y3}{output three}
  \item{y4}{output four}
}
}
\usage{
airline
}
\description{
A dataset containing 2 inputs and 4 outputs
}
\keyword{datasets}
