% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencyCEAT.R
\name{CEAT_BCC_in}
\alias{CEAT_BCC_in}
\title{Banker, Charnes and Cooper programming model with input orientation for a Convexified Efficiency Analysis Trees model}
\usage{
CEAT_BCC_in(j, scores, x_k, y_k, atreeTk, ytreeTk, nX, nY, N_leaves)
}
\arguments{
\item{j}{Number of DMUs.}

\item{scores}{\code{matrix}. Empty matrix for scores.}

\item{x_k}{\code{data.frame}. Set of input variables.}

\item{y_k}{\code{data.frame} Set of output variables.}

\item{atreeTk}{\code{matrix} Set of "a" Pareto-coordinates.}

\item{ytreeTk}{\code{matrix} Set of predictions.}

\item{nX}{Number of inputs.}

\item{nY}{Number of outputs.}

\item{N_leaves}{Number of leaf nodes.}
}
\value{
A numerical vector with scores.
}
\description{
Banker, Charnes and Cooper programming model with input orientation for a Convexified Efficiency Analysis Trees model.
}
