% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check Efficiency Analysis Tree.}
\usage{
check(tree)
}
\arguments{
\item{tree}{List structure with the tree nodes.}
}
\value{
Message indicating if the tree is acceptable or breaks some Pareto-dominance relationship.
}
\description{
This function verifies if tree keeps to Pareto-dominance properties.
}
