% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitEAT.R
\name{comparePareto}
\alias{comparePareto}
\title{Pareto-dominance relationships}
\usage{
comparePareto(t1, t2)
}
\arguments{
\item{t1}{A first node.}

\item{t2}{A second node.}
}
\value{
-1 if t1 dominates t2, 1 if t2 dominates t1 and 0 if there are no Pareto-dominance relationships.
}
\description{
This function denotes if a node dominates another one or if there is no Pareto-dominance relationship.
}
