% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compute_iters.R
\name{compute_iters}
\alias{compute_iters}
\title{Determine Number of Iterations to Simulate a Stochastic Model}
\usage{
compute_iters(abm, input_values, out, constraints = "none",
  parallel = FALSE, cores = NULL, verbose = TRUE,
  method = c("coef_var"))
}
\arguments{
\item{abm}{A function that takes as input values for each of the
\code{input_values}}

\item{input_values}{List}

\item{out}{Character vector length one to be passed an argument to the
\code{abm} function to specify what outcome measure to use.}

\item{constraints}{Optional Character vector that is either "none" or is using
only variable names that are specified in the input_values List argument.
This character vector is evaluated in an environment created for the sampled
data on the variables, and its evaluation results in a Logical vector that
that subsets sampled.}

\item{parallel}{Optional logical vector length one. Default is FALSE.}

\item{cores}{Optional Numeric vector length one. Default is
parallel::detectCores().}

\item{verbose}{Optional logical vector.}

\item{method}{Optional character vector that is either ...}
}
\value{
Returns a sensitivity object that can be plotted by functions.
}
\description{
\code{compute_iters} conducts estimates a sufficient number of iterations for
subsequent analysis of a simulation model.
}
\details{
This is function of the \strong{eat} package. It takes an abm in function form
and a list of input values.
}
\examples{
compute_iters()
}
\references{
Lorscheid, I., Heine, B.O., & Meyer, M. (2012). Opening the "black
box" of simulations: increased transparency and effective communication
through the systematic design of experiments. Computational and Mathematical
Organization Theory, 18 (1), 22–62.

Hendricks W, Robey K (1936) The sampling distribution of the coefficient of
variation. Ann Math Stat 7:129–132
}

