

\name{ConQuest.Log.Example1.log}

\docType{data}

\alias{ConQuest.Log.Example1.log}

\title{Example Log File from ConQuest}

\description{
This is a text file with the log from a ConQuest analysis
It can be accessed in path file.path( .Library , "eat/extdata/ConQuest.Log.Example1.log" )
}

%\usage{data(ConQuest.Log.Example1.log)}

\format{txt}

%\source{}

%\references{
%}

%\keyword{}






