% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_program.R
\name{create_fit_func}
\alias{create_fit_func}
\title{Create a fitness function to estimate a program (R function)}
\usage{
create_fit_func(loss_function, X, y, parallel)
}
\arguments{
\item{loss_function}{function resulting from a call to
\code{\link{create_loss_func}}.}

\item{X}{model.matrix}

\item{y}{outcome}
}
\value{
Returns a function that takes in a function as an argument and then
  assigns it a numeric fitness value we want to minimize. Use this in
  \code{\link{estimate_program}}.
}
\description{
Create a fitness function to estimate a program (R function)
}

