% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_program.R
\name{create_func_set}
\alias{create_func_set}
\title{Create function set}
\usage{
create_func_set(functions, link)
}
\arguments{
\item{functions}{Length one character vector}

\item{link}{Length one character vector to pass to create_link_func}
}
\value{
Returns a function set.
}
\description{
Create function set
}

