% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_program.R
\name{create_link_func}
\alias{create_link_func}
\title{Pick one of the link functions}
\usage{
create_link_func(link)
}
\arguments{
\item{link}{Character vector length one specifying which link function you
  want.}
}
\value{
Returns a function that has one argument (Numeric
  vector length one), returns a Numeric vector length one.
}
\description{
Pick one of the link functions
}
\examples{
create_link_func("logit")
}

