% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_sample.R
\name{create_set}
\alias{create_sample}
\alias{create_set}
\alias{keep_satisfied}
\title{Create set of samples by sampling with LHS and then checking constraints.}
\usage{
create_set(input_values, input_names, sample_count)

create_sample(input_values, input_names, sample_count)

keep_satisfied(sampled)
}
\arguments{
\item{input_values}{List}

\item{input_names}{Character vector}

\item{sample_count}{Numeric vector length one.}

\item{sampled}{Output of create sample_sample}
}
\value{
Returns a data.frame of samples.

Returns a data.frame of samples.

Returns a data.frame of samples.
}
\description{
\code{create_set} creates sample that stay within constraints.
}
\section{Functions}{
\itemize{
\item \code{create_sample}: Create a sample.

\item \code{keep_satisfied}: Stay within constraints.
}}

