% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eat_pkg.R
\docType{package}
\name{eat-package}
\alias{eat}
\alias{eat-package}
\title{EAT: Efficiency Analysis Trees}
\description{
This package provides functions to determine production frontiers and technical efficiency measures through non-parametric techniques based upon regression trees. The package includes code for estimating radial input, output, directional and additive measures, plotting graphical representations of the scores and the production frontiers by means of trees, and determining rankings of importance of input variables in the analysis. Additionally, an adaptation of Random Forest by a set of individual Efficiency Analysis Trees for estimating technical efficiency is also included.
}
\seealso{
Github: \url{https://github.com/MiriamEsteve/EAT} \cr
\cr
\insertRef{esteve2020}{eat}
}
\author{
\strong{Mantainers}: \itemize{
\item{Miriam Esteve Campello \email{miriam.estevec@umh.es}}
\item{Víctor Javier España Roch}
} \cr
Authors:
\itemize{
\item{Miriam Esteve Campello}
\item{Víctor Javier España Roch}
\item{Juan Aparicio Baeza}
\item{Xavier Barber Vallés}
}
}
\keyword{internal}
