% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eat.R
\docType{package}
\name{eat}
\alias{eat}
\alias{eat-package}
\title{eat: An R package for data-driven agent-based modeling}
\description{
\code{eat}'s main functions for estimating and testing the models:
\enumerate{
\item \code{\link{cv_abm}}
\item \code{\link{training}}
}
}

