% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencyEAT.R
\name{EAT_BCC_in}
\alias{EAT_BCC_in}
\title{Banker, Charnes and Cooper Programming Model with Input Orientation for an Efficiency Analysis Trees model}
\usage{
EAT_BCC_in(j, scores, x_k, y_k, atreeTk, ytreeTk, nX, nY, N_leaves)
}
\arguments{
\item{j}{Integer. Number of DMUs.}

\item{scores}{Matrix. Empty matrix for scores.}

\item{x_k}{Dataframe. Set of input variables.}

\item{y_k}{Dataframe. Set of output variables.}

\item{atreeTk}{Matrix. Set of "a" Pareto-coordinates.}

\item{ytreeTk}{Matrix. Set of predictions.}

\item{nX}{Integer. Number of inputs.}

\item{nY}{Integer. Number of outputs.}

\item{N_leaves}{Integer. Number of leaf nodes.}
}
\value{
A numerical vector with efficiency scores.
}
\description{
Banker, Charnes and Cooper programming model with input orientation for an Efficiency Analysis Tree model.
}
