% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{EAT_BCC_output}
\alias{EAT_BCC_output}
\title{BCC output-oriented}
\usage{
EAT_BCC_output(data, tree, x, y)
}
\arguments{
\item{data}{Data to be used.}

\item{tree}{Tree structure.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}
}
\value{
A numerical vector with efficiency scores.
}
\description{
BCC output-oriented
}
