% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{EAT_DDF}
\alias{EAT_DDF}
\title{EAT Directional Distance}
\usage{
EAT_DDF(data, tree, x, y)
}
\arguments{
\item{data}{Data to be used.}

\item{tree}{Tree structure.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}
}
\value{
A numerical vector with scores.
}
\description{
EAT Directional Distance
}
