% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_secuencial.R
\name{EAT_forest}
\alias{EAT_forest}
\title{Forest EAT}
\usage{
EAT_forest(data, x, y, numStop, k_forest)
}
\arguments{
\item{data}{Data to be used.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{numStop}{Minimun number of observations on a node to be splitted.}

\item{k_forest}{Number of variables chosen to do Random Forest.}
}
\value{
Deep tree from a certain subset of the original data.
}
\description{
This function generates a deep tree from each new subset from the original data.
}
