% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{EAT_frontier_levels}
\alias{EAT_frontier_levels}
\title{Output Levels in an Efficiency Analysis Trees model}
\usage{
EAT_frontier_levels(object)
}
\arguments{
\item{object}{An \code{EAT} object.}
}
\value{
A \code{data.frame} with the frontier output levels at the leaf nodes of the Efficiency Analysis Trees model introduced.
}
\description{
This function returns the frontier output levels for an Efficiency Analysis Trees model.
}
\examples{
simulated <- Y1.sim(N = 50, nX = 3)
EAT_model <- EAT(data = simulated, x = c(1, 2, 3), y = 4, numStop = 10, fold = 5)
EAT_frontier_levels(EAT_model)

}
