% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT_plot.R
\name{EAT_plot}
\alias{EAT_plot}
\title{Efficiency Analysis Trees Plot}
\usage{
EAT_plot(object)
}
\arguments{
\item{object}{An EAT object.}
}
\value{
Plot object with the following elements for each node:
\itemize{
\item{id}: node index .
\item{R}: mean square error in the node.
\item{Samples}: number of observations in the node.
\item{Variables}: splitting variable.
\item{y}: output prediction.
}
}
\description{
Plot an EAT model.
}
