% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{EAT_Rusell_input}
\alias{EAT_Rusell_input}
\title{Input oriented Rusell model}
\usage{
EAT_Rusell_input(data, tree, x, y)
}
\arguments{
\item{data}{Data to be used.}

\item{tree}{Tree structure.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}
}
\value{
A numerical vector with scores.
}
\description{
Input oriented Rusell model
}
