% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{EAT_size}
\alias{EAT_size}
\title{Number of Leaf Nodes in an Efficiency Analysis Trees model}
\usage{
EAT_size(object)
}
\arguments{
\item{object}{An \code{EAT} object.}
}
\value{
Number of leaf nodes of the Efficiency Analysis Trees model introduced.
}
\description{
This function returns the number of leaf nodes for an Efficiency Analysis Trees model.
}
\examples{
simulated <- Y1.sim(N = 50, nX = 3)
EAT_model <- EAT(data = simulated, x = c(1, 2, 3), y = 4, numStop = 10, fold = 5)
EAT_size(EAT_model)

}
