% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{eff.levels}
\alias{eff.levels}
\title{Efficiency Analysis Tree output efficiency levels}
\usage{
eff.levels(object)
}
\arguments{
\item{object}{An EAT object.}
}
\value{
Data frame with output efficiency levels.
}
\description{
This function returns the output efficiency levels of an Efficiency Analysis Tree model.
}
\examples{

simulated <- eat:::Y1.sim(N = 50, nX = 3)
model <- EAT(data = simulated, x = c(1, 2, 3), y = 4, numStop = 10, fold = 5)
eff.levels(model)

}
