% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{eff_data}
\alias{eff_data}
\title{Efficiecy Objects}
\usage{
eff_data(data, tree, x, y)
}
\arguments{
\item{data}{Data to be used.}

\item{tree}{Tree structure.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}
}
\value{
Number of rows, matrix for scoring, matrix of inputs, matrix of outputs, a Pareto-coordinates, predictions, number of inputs, number of outputs and number of leaf nodes.
}
\description{
This function prepares the necessary objects to get the scores.
}
