% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency_barplot}
\alias{efficiency_barplot}
\title{Efficiency Bar Plot}
\usage{
efficiency_barplot(scores, data, tree, x, y, color = T)
}
\arguments{
\item{scores}{Score object}

\item{data}{Data to be used.}

\item{tree}{Tree structure.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{color}{Logical. If True, observations with same efficient level output are painted the same level.}
}
\value{
Efficiency scores
}
\description{
This function returns a bar plot with scores
}
