% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency_density_barplot}
\alias{efficiency_density_barplot}
\title{Efficiency Density and Barplot}
\usage{
efficiency_density_barplot(scores, data, tree, x, y)
}
\arguments{
\item{scores}{Score object}

\item{data}{Data to be used.}

\item{tree}{Tree structure.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}
}
\value{
Efficiency scores
}
\description{
This function returns a density plot
}
