% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency_EAT}
\alias{efficiency_EAT}
\title{Efficiency Scores}
\usage{
efficiency_EAT(data, x, y, object, scores_model, r = 4, na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe for which the efficiency score is calculated.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{object}{An EAT object.}

\item{scores_model}{Mathematic programming model to calculate scores. 
\itemize{
\item{\code{EAT_BCC_out}} BBC model. Output orientation.
\item{\code{EAT_BCC_in}}  BBC model. Input orientation.
\item{\code{EAT_DDF}}     Directional distance model.
\item{\code{EAT_RSL_out}} Rusell model. Output orientation
\item{\code{EAT_RSL_in}}  Rusell model. Input orientation.
\item{\code{EAT_WAM}}     Weighted Additive model.
}}

\item{r}{Integer. Decimal units for scores.}

\item{na.rm}{Logical. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
Efficiency scores
}
\description{
This function calculates the efficiency scores for each DMU.
}
