% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency_FDH}
\alias{efficiency_FDH}
\title{FDH Efficiency Scores}
\usage{
efficiency_FDH(data, x, y, scores_model = "FDH_BCC_out", r = 4, na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe for which the efficiency score is calculated.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{scores_model}{Mathematic programming model to calculate scores. 
\itemize{
\item{\code{FDH_BCC_out}} BBC model. Output orientation.
\item{\code{FDH_BCC_in}}  BBC model. Input orientation.
\item{\code{FDH_DDF}}     Directional distance model.
\item{\code{FDH_RSL_out}} Rusell model. Output orientation
\item{\code{FDH_RSL_in}}  Rusell model. Input orientation.
\item{\code{FDH_WAM}}     Weighted Additive model.
}}

\item{r}{Integer. Decimal units for scores.}

\item{na.rm}{Logical. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
Efficiency scores
}
\description{
This function calculates the efficiency scores for each DMU by an FDH model.
}
