% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_plots.R
\name{efficiency_jitter}
\alias{efficiency_jitter}
\title{Efficiency Scores Jitter Plot}
\usage{
efficiency_jitter(object, scores_EAT, upb = 1, lwb = 0)
}
\arguments{
\item{object}{An EAT object.}

\item{scores_EAT}{Dataframe with scores.}

\item{upb}{Numeric. Upper bound for labeling.}

\item{lwb}{Numeric. Lower bound for labeling.}
}
\value{
Geom jitter plot with DMUs and scores.
}
\description{
This function returns a jitter plot. DMUs are grouped in nodes (same input paradigm) and its score are showed. A black point represents the score mean in a group and the line the standard deviation. Finally, the user can specify an upper bound (upb) and a lower bound (lwb) in order to show the labels.
}
