% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEAT.R
\name{efficiency_RFEAT}
\alias{efficiency_RFEAT}
\title{RFEAT Efficiency Scores}
\usage{
efficiency_RFEAT(data, x, y, object)
}
\arguments{
\item{data}{Dataframe for which the efficiency score is calculated.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{object}{A RFEAT object}
}
\value{
Dataframe with input variables and efficiency scores by a RFEAT model.
}
\description{
This function calculates the efficiency scores for each DMU by an RFEAT model.
}
