% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDH.R
\name{efficiencyFDH}
\alias{efficiencyFDH}
\title{FDH Efficiency Scores}
\usage{
efficiencyFDH(data, x, y, scores_model, r = 4, na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe for which the efficiency score is calculated.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{scores_model}{Mathematic programming model to calculate scores. 
\itemize{
\item{\code{FDH_BCC_out}} BBC model. Output orientation.
\item{\code{FDH_BCC_in}}  BBC model. Input orientation.
\item{\code{FDH_DDF}}     Directional distance model.
\item{\code{FDH_RSL_out}} Rusell model. Output orientation
\item{\code{FDH_RSL_in}}  Rusell model. Input orientation.
\item{\code{FDH_WAM}}     Weighted Additive model.
}}

\item{r}{Integer. Decimal units for scores.}

\item{na.rm}{Logical. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
Dataframe with input variables and efficiency scores by a FDH model.
}
\description{
This function calculates the efficiency scores for each DMU by an FDH model.
}
\examples{

X1 <- runif(50, 1, 10)
X2 <- runif(50, 2, 10)
Y1 <- log(X1) + 3 - abs(rnorm(50, mean = 0, sd = 0.4))
Y2 <- log(X1) + 2 - abs(rnorm(50, mean = 0, sd = 0.7))

simulated <- data.frame(x1 = X1, x2 = X2, y1 = Y1, y2 = Y2)

efficiencyFDH(data = simulated, x = c(1, 2), y = c(3, 4), scores_model = "FDH_BCC_out",
              r = 2, na.rm = TRUE)

}
