% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFEAT.R
\name{efficiencyRFEAT}
\alias{efficiencyRFEAT}
\title{RFEAT Efficiency Scores}
\usage{
efficiencyRFEAT(data, x, y, object)
}
\arguments{
\item{data}{Dataframe for which the efficiency score is calculated.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{object}{A RFEAT object}
}
\value{
Dataframe with input variables and efficiency scores by a RFEAT model.
}
\description{
This function calculates the efficiency scores for each DMU by an RFEAT model.
}
\examples{

X1 <- runif(50, 1, 10)
X2 <- runif(50, 2, 10)
Y1 <- log(X1) + 3 - abs(rnorm(50, mean = 0, sd = 0.4))
Y2 <- log(X1) + 2 - abs(rnorm(50, mean = 0, sd = 0.7))

simulated <- data.frame(x1 = X1, x2 = X2, y1 = Y1, y2 = Y2)

n <- nrow(simulated)
t_index <- sample(1:n, n * 0.8)
training <- simulated[t_index, ]
test <- simulated[-t_index, ]

RFEAT_model <- RFEAT(data = training, x = c(1,2), y = c(3, 4), numStop = 5,
                     m = 50, s_mtry = "Breiman", na.rm = TRUE)

efficiencyRFEAT(data = training, x = c(1, 2), y = c(3, 4), object = RFEAT_model)

efficiencyRFEAT(data = test, x = c(1, 2), y = c(3, 4), object = RFEAT_model)

}
