% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_abm.R
\name{estimate_abm}
\alias{estimate_abm}
\title{Estimate an ABM}
\usage{
estimate_abm(data, features, Formula, agg_patterns, abm_simulate, abm_vars,
  iters, tseries_len, verbose = TRUE, tp = rep(tseries_len,
  nrow(agg_patterns)), package = c("caretglm", "caretglmnet", "glm",
  "caretnnet", "caretdnn"), sampling = FALSE, sampling_size = 1000,
  STAT = c("mean", "median"), abm_optim = c("GA", "DE"),
  optimize_abm_par = FALSE, parallel_training = FALSE)
}
\arguments{
\item{data}{\code{data.frame} with each row (obervational unit) being an 
individual decision. With a column named "group" specifying which group of 
\code{agg_patterns} each obseravtion is in, and a column named "period" 
specifying at what time period each behavior was taken.}

\item{features}{\code{list} of the variables (columns in \code{data}) to be 
used in the prediction \code{Formula}. As many elements in the \code{list} 
as we want discrete models for different times. Each element of the 
\code{list} is a \code{character vector}, with each element of the 
\code{character vector} being a feature to use for training an 
individual-level model.}

\item{Formula}{\code{list} where each element is a length one character vector
that specifies a formula, e.g. \code{"y ~ x"}. The character vector makes 
sense in the context of the \code{features} and \code{data}. There are as 
many elements in the list as there are discrete models for different times.}

\item{agg_patterns}{data.frame with rows (observational unit) being the group 
and columns: (a.) those aggregate level variables needed for the prediction 
with the specified \code{formula} (with same names as the variables in the 
formula); (b.) a column named "action" with the proportion of the relevant 
outcome action taken in that group; (c.) columns named 
\code{paste(seq(tseries_len))} with the mean/median levels (\code{STAT}) of 
the action for each time period.}

\item{abm_simulate}{function with these arguments: \code{model, features, 
parameters, tuning_parameters, iterations, time_len, STAT = c("mean", 
"median")}. Where \code{model} is the output of \code{\link{training}}. 
Output of the function is a list with three named elements: \code{dynamics, 
action_avg, simdata}. Where \code{dynamics} is a numeric vector length 
\code{tseries_len}, \code{action_avg} is a numeric vector length one, and 
\code{simdata} is a \code{data.frame} with the numeric results of the 
simulation.}

\item{abm_vars}{a list with either (1.) a numeric vector named "lower" AND a 
numeric vector named "upper" each the length of the number of tuning_params 
of ABM (the names of the elements of these vecs should be the names of the 
variables and they should be in the same order that the \code{abm_simulate} 
function uses them); or (2.) a numeric vector named "value" the length of 
the number of tuning_params of the ABM (variables should be in the same 
order that the \code{abm_simulate} function uses them). Either provide lower
and upper elements of the list or provide a value element of the list.}

\item{iters}{numeric vector length one specifying number of iterations to 
simulate ABM for.}

\item{tseries_len}{numeric vector length one specifying maximum number of time
periods to use for model training and testing. If some groups have less than
the maximum then you need to provide a vector to the \code{tp} argument.}

\item{verbose}{optional logical vector length one, default is \code{TRUE}.}

\item{tp}{optional numeric vector length number of rows of \code{agg_patterns}
specifying how long the time series for each group should be. Default is 
\code{rep(tseries_len, nrow(agg_patterns))}.}

\item{package}{optional character vector length one, default is 
\code{"caretglm", "caretglmnet", "glm", "caretnnet", "caretdnn"}.}

\item{sampling}{optional logical vector length one, default is \code{FALSE}. 
If \code{sampling == TRUE}, we sample equal numbers of observations from 
each 'group' to reduce potential problems with the final estimated model 
being too affected by groups with more observations.}

\item{sampling_size}{optional numeric vector length one specifying how many 
observations from each group that \code{\link{training}} should sample to 
train the model, default is 1000. Only applicable when \code{sampling} 
argument is set to \code{TRUE}.}

\item{STAT}{optional character vector length one, default is \code{c("mean", 
"median")}.}

\item{abm_optim}{optional character vector length one, default is 
\code{c("GA", "DE")}.}

\item{optimize_abm_par}{optional logical vector length one, default is 
\code{FALSE}. This is passed to the optimization algorithm.}

\item{parallel_training}{optional logical vector length one, default is 
\code{FALSE}. This is passed to \code{\link{training}}.}
}
\value{
Returns a function that has three arguments: \code{parameters, out, 
 iterations}. If \code{out=="action_avg"} for the returned function, the 
 average of all the actions is returned by this function; otherwise, the 
 vector of the average for each time is returned by this function. This 
 returns a wrapper function around their abm simulation function to be used 
 for analysis.
}
\description{
Using \code{\link{estimate_abm}} one can use their data and their abm function
they are using for \code{\link{cv_abm}} to estimate an ABM via optimization of
its global abm parameters or their specification. Then this can be used for 
analysis.
}

