% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_program.R
\name{estimate_program}
\alias{estimate_program}
\title{Estimate a program (R function) from data.}
\usage{
estimate_program(formula, data, subset = NULL, loss = c("log_lik", "rmse",
  "identity", "identity_multi_class"), link = c("logit", "probit", "cauchit",
  "identity"), functions = c("math", "logical", "randomness"), mins = 2,
  parallel = FALSE, cores = NULL)
}
\arguments{
\item{formula}{formula used to create the data.frame needed, ensuring that the
outcome variable, the variable to the left of "~", is the first column in
the data.frame and the following columns are the predictor variables.}

\item{data}{a data.frame with named columns containing the variables in
formula. Neither a matrix nor an array will be accepted. We use the
\code{formula} to turn this into a data.frame where (i.) the first column is
named \code{"outcome"} and has the outcome variable we are evolving programs
to predict; (ii.) all other columns (there must be at least one other
column) that follow the \code{"outcome"} column are named columns containing
the predictor variables, which can be of any type. The order matters: when
using the evolved program for predictions the (named) arguments to the
function will be the predictor variables in the order they are supplied to
this \code{estimate_program()} function.}

\item{subset}{a specification of the rows to be used: defaults to all rows.
This can be any valid indexing vector (see [.data.frame) for the rows of
data or if that is not supplied, a data frame made up of the variables used
in formula.}

\item{loss}{Optional Character vector length one}

\item{link}{Optional Character vector length one}

\item{functions}{Optional Character vector as long as there are (prespecified)
categories (\code{c("math", "logical", "randomness")}) of functions that you
want to include}

\item{mins}{Optional Integer vector length one}

\item{parallel}{Optional Logical vector length one. Default is \code{parallel
= FALSE}; \code{parallel = TRUE} can be slower if the data set is small
relative to the numner of population evolutions desired}

\item{cores}{Optional Integer vector length one}
}
\value{
The function returns an S4 object. See \linkS4class{estimate_program}
 for the details of the \code{slots} (objects) that this type of object will
 have.
}
\description{
\code{estimate_program} uses \code{rgp}'s \code{rgp::typedGeneticProgramming}.
}

