% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitEAT.R
\name{estimEAT}
\alias{estimEAT}
\title{Estimation of child nodes}
\usage{
estimEAT(data, leaves, t, xi, s, y)
}
\arguments{
\item{data}{Data to be used.}

\item{leaves}{List structure with leaf nodes or pending expansion nodes.}

\item{t}{Node which is being split.}

\item{xi}{Variable index that produces the split.}

\item{s}{Value of xi variable that produces the split.}

\item{y}{Column output indexes in data.}
}
\value{
Left and right children nodes.
}
\description{
This function gets the estimation of the response variable and updates Pareto-coordinates and the observation index for both new nodes.
}
