% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDH.R
\name{FDH}
\alias{FDH}
\title{FDH: Free Disposal Hull}
\usage{
FDH(data, x, y)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}
}
\value{
List structure with the train and the test set.
}
\description{
This function generates a FDH model
}
