% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_secuencial.R
\name{forest_scoreEAT}
\alias{forest_scoreEAT}
\title{Forest Score EAT}
\usage{
forest_scoreEAT(forest, data, x, y, nameCol = "Score")
}
\arguments{
\item{forest}{Random Forest EAT object.}

\item{data}{Data to be used.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{nameCol}{Name of coulumn}
}
\value{
...
}
\description{
This function ...
}
