% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontier.R
\name{frontier}
\alias{frontier}
\title{EAT frontier}
\usage{
frontier(data, tree, x, y, rownames = FALSE)
}
\arguments{
\item{data}{Data used for modeling.}

\item{tree}{List structure with the tree nodes.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{rownames}{logical. If T, row names are displayed in the plot instead of points.}
}
\value{
Scatter plot with DMUs and frontier predicted by `EAT`
}
\description{
This function creates a plot with EAT frontier for the scenario corresponding to one input and one output.
}
