% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesForRCV.R
\name{generateLv}
\alias{generateLv}
\title{Train and test sets generation.}
\usage{
generateLv(data, fold)
}
\arguments{
\item{data}{Data to be splited in train and test.}

\item{fold}{Parts in which is divided the original set to do Cross-Validation.}
}
\value{
List structure with the train and the test set.
}
\description{
This function split the original data in two new data sets: a train set and a test set.
}
