% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{imp_var_EAT}
\alias{imp_var_EAT}
\title{Importance variable of Breiman}
\usage{
imp_var_EAT(data, tree, x, y, r)
}
\arguments{
\item{data}{Data from EAT object.}

\item{tree}{Tree from EAT object.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{r}{Decimal units.}
}
\value{
A dataframe with the best split for each node and variable and its importance.
}
\description{
This function recalculates all the possible splits, with the exception of the actual used, and for each node and variable gets the best split based on their degree of importance.
}
