% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingRFEAT.R
\name{imp_var_RFEAT}
\alias{imp_var_RFEAT}
\title{Importance variable through Random Forest + Efficiency Analysis Trees}
\usage{
imp_var_RFEAT(object, r = 2)
}
\arguments{
\item{object}{A RFEAT object}

\item{r}{Integer. Decimal units.}
}
\value{
Vector of input importance scores
}
\description{
Importance variable through Random Forest + Efficiency Analysis Trees.
}
