% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{M_Breiman}
\alias{M_Breiman}
\title{Breiman Importance}
\usage{
M_Breiman(object, r)
}
\arguments{
\item{object}{An EAT or a RFEAT object.}

\item{r}{Decimal units.}
}
\value{
Dataframe with one column and the importance of each variable in rows.
}
\description{
This function evaluates the importance of each predictor by the notion of surrogate splits.
}
