% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitRFEAT.R
\name{mtry_inputSelection}
\alias{mtry_inputSelection}
\title{Random Selection of Variables}
\usage{
mtry_inputSelection(data, x, t, mtry)
}
\arguments{
\item{data}{Dataframe containing the training set.}

\item{x}{Vector. Column input indexes in data.}

\item{t}{Node which is being split.}

\item{mtry}{Integer. Number of inputs selected for a node to be split.}
}
\value{
Index of the variables by which the node is divided.
}
\description{
This function randomly selects the variables that are evaluated to divide a node and removes those that do not present variability.
}
