% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_def_cv_abm.R
\name{performance}
\alias{performance}
\title{Turns cv_abm S4 object into useful outputs for summary or plotting.}
\usage{
performance(results, output = c("MSE", "cor", "cor_pval", "SE", "plot"))
}
\arguments{
\item{results}{S4 cv_abm object}

\item{output}{Optional character vector length one indicating what the
  desired output is, must be one of: \code{c("MSE", "cor", "cor_pval", "SE",
  "plot")}.}
}
\description{
Turns cv_abm S4 object into useful outputs for summary or plotting.
}

