% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pc_sa_plot.R
\name{plot_pc}
\alias{plot_pc}
\title{Plot Partial Correlation Analysis of a Simulation Model}
\usage{
plot_pc(x, outcome_var = "Outcome", xlab = "Variable")
}
\arguments{
\item{x}{An object created by \code{pc_sa}.}

\item{outcome_var}{Optional character vector for labeling the outcome variable
in the plot. Default is "Outcome".}

\item{xlab}{Optional character vector for labeling the variables.}
}
\value{
Returns a ggplot2 plot.
}
\description{
\code{plot_pc}
}
\details{
This is function of the \strong{eat} package. \code{pc_sa} conducts a
a partial correlation analysis.
}
\examples{
fake_abm <- function(params, out) {
  x1 <- params[1]
  x2 <- params[2]
  if (out=="sq") return(x1^2 + x2 + rnorm(1, 0))
  if (out=="ident") return(x1 + x2 + rnorm(1, 0))
}
inputs <- lapply(list(param1 = NA, param2 = NA),
                 function(x) list(random_function = "qunif",
                                  ARGS = list(min = 0, max = 1)))
s <- pc_sa(fake_abm, inputs, "sq")
# plot_pc(s)
}

