% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sobol_sa_plot.R
\name{plot_sobol_fo}
\alias{plot_sobol_fo}
\title{Plot First-Order Effects from a Sobol Sensitivity Analysis of a Simulation
Model}
\usage{
plot_sobol_fo(x, outcome_var = "Outcome")
}
\arguments{
\item{x}{An object created by \code{sobol_sa}. \code{input_values}}

\item{outcome_var}{Optional character vector for labeling the outcome variable
 in the plot. Default is "Outcome".}
}
\value{
Returns a ggplot2 plot.
}
\description{
\code{plot_sobol_fo} plots first-order effects.
}
\details{
This is function of the \strong{eat} package. \code{sobol_sa} conducts a
global variance decomposition, and then \code{plot_sobol_fo} can be used to
plot it.
}
\examples{
fake_abm <- function(params, out) {
  x1 <- params[1]
  x2 <- params[2]
  if (out=="sq") return(x1^2 + x2 + rnorm(1, 0))
  if (out=="ident") return(x1 + x2 + rnorm(1, 0))
}
inputs <- lapply(list(param1 = NA, param2 = NA),
                 function(x) list(random_function = "qunif",
                                  ARGS = list(min = 0, max = 1)))
s <- sobol_sa(fake_abm, inputs, "sq")
# plot_sobol_fo(s)
}

