% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{posIdNode}
\alias{posIdNode}
\title{Position of the node}
\usage{
posIdNode(tree, idNode)
}
\arguments{
\item{tree}{List structure with the tree nodes.}

\item{idNode}{Id of a specific node.}
}
\value{
Position of the node or -1 if it is not found.
}
\description{
This function looks for the position of a given node in the tree.
}
