% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predict}
\alias{predict}
\title{Predict}
\usage{
predict(tree, data, x, y)
}
\arguments{
\item{tree}{List structure with the tree nodes.}

\item{data}{Data to be predicted.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}
}
\value{
A data frame with the expected value in one column for each dependent variable based on the given dataset.
}
\description{
This function predicts the expected value based on a dataset
}
