% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDH.R
\name{predict_FDH}
\alias{predict_FDH}
\title{Model prediction for FDH}
\usage{
predict_FDH(data, x, y)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}
}
\value{
Data frame with the original data and the predicted values by and FDH model.
}
\description{
This function predicts the expected output by an FDH model.
}
