% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predictor}
\alias{predictor}
\title{Predictor}
\usage{
predictor(tree, register)
}
\arguments{
\item{tree}{List structure with the tree nodes.}

\item{register}{Set of independent values.}
}
\value{
The expected value of the dependent variable based on the given register.
}
\description{
This function predicts the expected value based on a set of inputs.
}
