% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFEAT.R
\name{predictRFEAT}
\alias{predictRFEAT}
\title{Model prediction for RFEAT}
\usage{
predictRFEAT(object, newdata)
}
\arguments{
\item{object}{A RFEAT object.}

\item{newdata}{Dataframe. Set of input variables to predict on.}
}
\value{
Data frame with the original data and the predicted values.
}
\description{
This function predicts the expected output by an RFEAT object.
}
