% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preProcess}
\alias{preProcess}
\title{Data errors}
\usage{
preProcess(data, x, y, na.rm = T, check = 0)
}
\arguments{
\item{data}{Data to be used. Dataframe or matrix objects are acceptable.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{na.rm}{If True, NA rows are omitted.}

\item{check}{If check = 1, the function comes from predictor function and thus NA's should not be omitted.}
}
\value{
Data processed
}
\description{
This function displays error messages in relation with data such as presence of NA or not allowed classes of data.
}
