% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{print_results}
\alias{print_results}
\title{Print leaf nodes results}
\usage{
print_results(EAT)
}
\arguments{
\item{EAT}{A EAT object}
}
\value{
Printing in table format with the data described above.
}
\description{
This function prints the number of observations and the proportion, the efficiency level of the outputs and the mean square error for each final node.
}
