% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_secuencial.R
\name{random_forest_secuencial}
\alias{random_forest_secuencial}
\title{Random Forest Secuencial}
\usage{
random_forest_secuencial(m, data, x, y, numStop = 5, k_forest = 2)
}
\arguments{
\item{m}{Number of trees.}

\item{data}{Data to be used.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{numStop}{Minimun number of observations on a node to be splitted.}

\item{k_forest}{Number of variables chosen to do Random Forest.}
}
\value{
List with m tree models.
}
\description{
This function generates a set of tree models with different samples of the original data.
}
