% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_secuencial.R
\name{random_histogram}
\alias{random_histogram}
\title{Histogram of Y}
\usage{
random_histogram(data, nameCol = "Score")
}
\arguments{
\item{data}{Data to be used with scores.}

\item{nameCol}{Name of coulumn}
}
\value{
...
}
\description{
This function ...
}
