% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_secuencial.R
\name{random_predictor}
\alias{random_predictor}
\title{Random Forest Predictor}
\usage{
random_predictor(forest, Xn)
}
\arguments{
\item{forest}{Random Forest EAT object.}
}
\value{
List with m tree models.
}
\description{
This function predicts ...
}
