% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_secuencial.R
\name{random_scoreEAT}
\alias{random_scoreEAT}
\title{Random Score EAT}
\usage{
random_scoreEAT(data, tree, x, y, nameCol)
}
\arguments{
\item{data}{Data to be used.}

\item{tree}{Tree of forest object}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{nameCol}{Name of coulumn}
}
\value{
...
}
\description{
This function ...
}
