% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_secuencial.R
\name{random_test}
\alias{random_test}
\title{Random Test}
\usage{
random_test(forest, df_test)
}
\arguments{
\item{forest}{Random Forest EAT object.}

\item{df_test}{Data to be predicted.}
}
\value{
...
}
\description{
This function predicts the output of a independent variables data set.
}
