% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEAT.R
\name{RandomEAT}
\alias{RandomEAT}
\title{Random Forest EAT}
\usage{
RandomEAT(data, x, y, numStop, s_mtry)
}
\arguments{
\item{data}{Data to be used.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in}

\item{numStop}{Ending rule}

\item{s_mtry}{Select number of inputs. It could be: "Breiman", "DEA1", "DEA2", "DEA3" or "DEA4"}
}
\value{
List of m trees in forest and the error that will be used in the ranking of the importance of the variables.
}
\description{
This function is a middle function to build "m" EAT trees in forest
}
