% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking}
\alias{ranking}
\title{Ranking of variables}
\usage{
ranking(object, r = 2, threshold = 70, barplot = TRUE)
}
\arguments{
\item{object}{An EAT or a RFEAT object.}

\item{r}{Integer. Decimal units.}

\item{threshold}{Numeric. Importance score value in which a broken line should be graphed.}

\item{barplot}{Logical. If True, barplot with importance scores is displayed.}
}
\value{

}
\description{
This function calculates variable importance for an EAT or RFEAT object.
}
