% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEAT.R
\name{ranking_RFEAT}
\alias{ranking_RFEAT}
\title{Ranking of variables by RFEAT}
\usage{
ranking_RFEAT(object, r = 2, barplot = TRUE)
}
\arguments{
\item{object}{An RFEAT object.}

\item{r}{Integer. Decimal units.}

\item{barplot}{Logical. If \code{TRUE}, a barplot with importance scores is displayed.}
}
\value{
Dataframe with scores or list with scores and barplot.
}
\description{
This function calculates variable importance for an RFEAT object.
}
