% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{RBranch}
\alias{RBranch}
\title{Error and leaves of the branch to be pruned}
\usage{
RBranch(t, tree)
}
\arguments{
\item{t}{List. A given EAT node.}

\item{tree}{A list containing EAT nodes.}
}
\value{
List with the sum of the errors of the children of the given node and the total number of leaf nodes that comes from it.
}
\description{
This function calculates the error of a branch as the sum of the errors of its children and the total number of leaf nodes that come from it.
}
