% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruningScoring.R
\name{RCV}
\alias{RCV}
\title{RCV}
\usage{
RCV(N, Lv, y, alphaIprim, fold, TAiv)
}
\arguments{
\item{N}{Number of rows in data.}

\item{Lv}{Test set.}

\item{y}{Column output indexes in data.}

\item{alphaIprim}{Alpha obtained as the square root of the product of two consecutive alpha values in tree_alpha list. It is used to find the best pruning tree.}

\item{fold}{Parts in which the original data is divided into to perform Cross-Validation.}

\item{TAiv}{List with each possible pruning for the deep tree generated with the train set and its associated alpha values.}
}
\value{
Set of best pruning and the associated error calculated with test sets.
}
\description{
RCV
}
