% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFEAT.R
\name{RF_predictor}
\alias{RF_predictor}
\title{Random Forest EAT predictions}
\usage{
RF_predictor(forest, xn)
}
\arguments{
\item{forest}{Trees structures in forest list.}

\item{xn}{Row indexes in data.}
}
\value{
Number of rows, matrix of inputs, matrix of outputs, predictions, number of inputs, number of outputs and number of leaf nodes.
}
\description{
Efficiency score of Random Forest EAT
}
