% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFEAT.R
\name{RFEAT}
\alias{RFEAT}
\title{Random Forest EAT}
\usage{
RFEAT(data, x, y, numStop = 5, m = 50, s_mtry = "Breiman", na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{numStop}{Integer. Minimun number of observations in a node for a split to be attempted.}

\item{m}{Integer. Number of trees to be build.}

\item{s_mtry}{Select number of inputs in each split. It can be an integer or any of the following options:
\itemize{
\item{\code{"Breiman"}}: \code{in / 3}
\item{\code{"DEA1"}}: \code{(obs / 2) - out}  
\item{\code{"DEA2"}}: \code{(obs / 3) - out}
\item{\code{"DEA3"}}: \code{obs - 2 * out}
\item{\code{"DEA4"}}: \code{min(obs / out, (obs / 3) - out)}
}}

\item{na.rm}{Logical. If \code{TRUE}, NA rows are omitted.}
}
\value{
A RFEAT object.
}
\description{
This function build \code{m} individual Efficiency Analysis Trees in a forest structure.
}
\examples{

X1 <- runif(50, 1, 10)
X2 <- runif(50, 2, 10)
Y1 <- log(X1) + 3 - abs(rnorm(50, mean = 0, sd = 0.4))
Y2 <- log(X1) + 2 - abs(rnorm(50, mean = 0, sd = 0.7))

simulated <- data.frame(x1 = X1, x2 = X2, y1 = Y1, y2 = Y2)

Rf_model <- RFEAT(data = simulated, x = c(1,2), y = c(3, 4), numStop = 5,
                  m = 100, s_mtry = "Breiman", na.rm = TRUE)

}
