

\name{science1.testlets}

\docType{data}

\alias{science1.testlets}

\title{Science achievement test data - Testlet definition as dummy codes}

\description{
This data frame contains testlet definitions as dummy codes for usage with \code{\link{automateModels}} and data set \code{\link{science1}}
}

%\usage{ }

\format{'data.frame':   185 obs. of  53 variables}

%\source{}

%\references{
%}

\keyword{datasets}






