

\name{science2}

\docType{data}

\alias{science2}

\title{Reading and writing achievement test data}

\description{
This data set contains fictional achievemetn scores of 4619 students of 3 countries. See format for a short description 
of variables.
}

\usage{data(science2)}

\format{'data.frame':   4619 obs. of  24 variables
  \describe{
    \item{idstud}{Unique identifier}
    \item{wgtSTUD}{variable of individual student weights}
    \item{sex}{Examinee's sex}
    \item{country}{Country where the examinee is from.}
    \item{JKZone}{jackknifing zone}
    \item{JKrep}{replicate ID}
    \item{reading_score1}{First plausible value of the reading score}
    \item{reading_score2}{Second plausible value of the reading score}
    \item{reading_score3}{Third plausible value of the reading score}
    \item{writing_score1}{First plausible value of the writing score}
    \item{writing_score2}{Second plausible value of the writing score}
    \item{writing_score3}{Third plausible value of the writing score}
	\item{passed_reading1}{First indicator whether examinee passed the reading minimal requirement}
	\item{passed_reading2}{Second indicator whether examinee passed the reading minimal requirement}
	\item{passed_reading3}{Third indicator whether examinee passed the reading minimal requirement}
	\item{passed_writing1}{First indicator whether examinee passed the writing minimal requirement}
	\item{passed_writing2}{Second indicator whether examinee passed the writing minimal requirement}
	\item{passed_writing3}{Third indicator whether examinee passed the writing minimal requirement}
	\item{zehisei}{Overall five indicators of highest socio-economic status. Think of it as imputed values. }
	\item{income}{Overall two indicators of mean month income.}
 }
}

\source{Simulated data}

%\references{
%}

\keyword{datasets}


