% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruningScoring.R
\name{scores}
\alias{scores}
\title{Pruning Scores}
\usage{
scores(N, Lv_notLv, x, y, fold, numStop, Tk, tree_alpha_list)
}
\arguments{
\item{N}{Number of rows in data.}

\item{Lv_notLv}{List with train and test sets.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{fold}{Parts in which the original data set is divided to perform Cross-Validation.}

\item{numStop}{Minimum number of observations on a node to be split.}

\item{Tk}{Best pruned tree.}

\item{tree_alpha_list}{List with all the possible pruning and its associated alpha.}
}
\value{
List with the best pruning for each fold, the pruning with a lower score and tree_alpha_list with scores updated.
}
\description{
This function calculates the score for each pruning of tree_alpha_list.
}
