% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEAT.R
\name{select_mtry}
\alias{select_mtry}
\title{Select m_try}
\usage{
select_mtry(s_mtry, t, nX, nY)
}
\arguments{
\item{s_mtry}{Select number of inputs. It could be: "Breiman", "DEA1", "DEA2", "DEA3" or "DEA4"}

\item{t}{Node}

\item{nX}{Number of columns input indexes in data.}

\item{nY}{Number of columns output indexes in}
}
\value{
Number of inputs selected according with the rule specified.
}
\description{
This function is a middle function to build "m" EAT trees in forest
}
