% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTk.R
\name{selectTk}
\alias{selectTk}
\title{Select Tk}
\usage{
selectTk(Tk, tree_alpha_list, SE)
}
\arguments{
\item{Tk}{Best pruned tree score.}

\item{tree_alpha_list}{List with all the possible pruning and its associated alpha and scores.}

\item{SE}{Value to get a range where new prunings is found.}
}
\value{
The same best tree or a new suitable one.
}
\description{
This function tries to find a new pruned tree with a shorter length and a score in the range generated for SE.
}
