% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTk.R
\name{SERules}
\alias{SERules}
\title{SERules}
\usage{
SERules(N, Lv, y, fold, Tk_score, BestTivs)
}
\arguments{
\item{N}{Number of rows in data.}

\item{Lv}{Test set.}

\item{y}{Column output indexes in data.}

\item{fold}{Parts in which the original data set is divided to perform Cross-Validation.}

\item{Tk_score}{Best pruned tree score.}

\item{BestTivs}{List of best pruned trees for each training set.}
}
\value{
Value to get a range where new pruning is found.
}
\description{
Based on Validation tests over BestTivs, a new range of scores is obtained to find new pruned trees.
}
