% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tree.R
\name{show_tree}
\alias{show_tree}
\title{Print tree}
\usage{
show_tree(tree)
}
\arguments{
\item{tree}{List structure with tree nodes.}
}
\value{
Data frame with the values of a scenario with nX inputs and nY outputs.
}
\description{
This function is used to simulate the data in a multioutput scenario.
}
