% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\name{simulated_data}
\alias{simulated_data}
\title{Multioutput data generation}
\usage{
simulated_data(N, nX, nY, border, noise = NULL)
}
\arguments{
\item{N}{Sample size.}

\item{nX}{Number of inputs.}

\item{nY}{Number of outputs.}

\item{border}{Percentage of DMU's in the frontier.}

\item{noise}{Random noise.}
}
\value{
Data frame with the values of a scenario with nX inputs and nY outputs.
}
\description{
This function is used to simulate the data in a multioutput scenario.
}
