% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{size}
\alias{size}
\title{EAT size}
\usage{
size(object)
}
\arguments{
\item{object}{An EAT object.}
}
\value{
Number of leaf nodes in the tree
}
\description{
This function calculates the number of leaf nodes in the tree.
}
