% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitEAT.R
\name{split}
\alias{split}
\title{Split node}
\usage{
split(data, tree, leaves, t, x, y, numStop)
}
\arguments{
\item{data}{Data to be used.}

\item{tree}{List structure with the tree nodes.}

\item{leaves}{List with leaf nodes or pending expansion nodes.}

\item{t}{Node which is being split.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{numStop}{Minimum number of observations in a node to be split.}
}
\value{
Leaves and tree lists updated with the new child nodes.
}
\description{
This function gets the variable and split value to be used in estimEAT, selects the best split and updates VarInfo, node indexes and leaves list.
}
