% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_secuencial.R
\name{split_forest}
\alias{split_forest}
\title{Split forest}
\usage{
split_forest(data, tree, leaves, t, x, y, numStop, arrayK)
}
\arguments{
\item{data}{Data to be used.}

\item{tree}{List structure with the tree nodes.}

\item{leaves}{List with leaf nodes or pending expansion nodes.}

\item{t}{Node which is being splitted.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{numStop}{Minimun number of observations on a node to be splitted.}

\item{arrayK}{Column input indexes to split.}
}
\value{
Leaves and tree lists updated with the new children nodes.
}
\description{
This function gets the variable and split value to be used in estimEAT, selects the best split and updates VarInfo, node indexes and leaves list.
}
