% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/training.R
\name{training}
\alias{training}
\title{Estimate an Agent Model}
\usage{
training(trainData, features, Formula, k, sampling = FALSE,
  sampling_size = 1000, outcome_var_name = "my.decision",
  package = c("caretglm", "caretglmnet", "glm", "caretnnet", "caretdnn"),
  parallel = FALSE, cv_type = c("cv", "repeatedcv"))
}
\arguments{
\item{trainData}{data.frame with each row (obervational unit) being an individual decision. With
a column called "group" specifying which group of \code{agg_patterns} each obseravtion is in.}

\item{features}{list of the variables (columns in \code{data}) to be used in the prediction
\code{Formula}. As many elements in the list as we want discrete models for
different times. Each element of the list is a character vector, with each element of the
character vector being a feature to use for training an individual-level model.}

\item{Formula}{list where each element is a length one character vector that specifies
a formula, e.g. \code{y ~ x}. The character vector makes sense in the context of the
\code{features} and \code{data}. There are as many elements in the list as there are
discrete models for different times.}

\item{k}{numeric vector length one specifying the number of models for different times.}

\item{sampling}{optional logical vector length one, default is \code{FALSE}.}

\item{sampling_size}{optional numeric vector length one specifying how many observations
from each group that \code{\link{training}} should sample to train the model, default is 1000.
Only applicable when \code{sampling} argument is set to \code{TRUE}.}

\item{outcome_var_name}{optional character vector length one, default is \code{"my.decision"}.
\code{\link{training}} uses it to sample to train the model with a balanced sampling based
on \code{outcome_var_name}. Only applicable when \code{sampling} argument is set to \code{TRUE}.}

\item{package}{optional character vector length one, default is
\code{"caretglm", "caretglmnet", "glm", "caretnnet", "caretdnn"}.}

\item{parallel}{optional logical vector length one, default is \code{FALSE}.}

\item{cv_type}{optional character vector length one, default is \code{c("cv", "repeatedcv")}.}
}
\value{
Returns a \code{list} length \code{k}.
}
\description{
\code{training} uses a \code{caret::train} or \code{stats::glm} algorithm to estimate an
individual-level model
}

