% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{two.out_two.in_scenario}
\alias{two.out_two.in_scenario}
\title{2 inputs & 2 outputs scenario}
\usage{
two.out_two.in_scenario(N, border, noise = NULL)
}
\arguments{
\item{N}{Sample size.}

\item{border}{Percentage of DMU's in the frontier.}

\item{noise}{Random noise.}
}
\value{
Data frame with the values of a scenario with nX inputs and nY outputs.
}
\description{
This function is used to simulate the data in a scenario with 2 inputs and 2 outputs.
}
