% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{X2Y2.simul}
\alias{X2Y2.simul}
\title{2 inputs & 2 outputs scenario}
\usage{
X2Y2.simul(N, border, noise = NULL)
}
\arguments{
\item{N}{Sample size.}

\item{border}{Percentage of DMU's in the frontier.}

\item{noise}{Random noise.}
}
\value{
Data frame with the values of a scenario with 2 inputs and 2 outputs.
}
\description{
This function is used to simulate the data in a scenario with 2 inputs and 2 outputs.
}
