% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{yFDH}
\alias{yFDH}
\title{Free Disposal Hull predictor}
\usage{
yFDH(data, x, y)
}
\arguments{
\item{data}{Data frame or matrix.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}
}
\description{
...
}
