% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\title{Alpha}
\usage{
alpha(tree)
}
\arguments{
\item{tree}{List structure with the tree nodes.}
}
\value{
Minimun alpha associated with a suitable node to be pruned.
}
\description{
This function gets the minimun alpha for each possible pruning.
}
