% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{alpha}
\alias{alpha}
\title{Alpha Calculation for Pruning Procedure}
\usage{
alpha(tree)
}
\arguments{
\item{tree}{A list containing EAT nodes.}
}
\value{
Numeric value corresponding to the minimun alpha associated with a suitable node to be pruned.
}
\description{
This function gets the minimun alpha for each subtree evaluated during the pruning procedure of Efficiency Analysis Trees.
}
