% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitRFEAT.R
\name{bagging}
\alias{bagging}
\title{Bagging data}
\usage{
bagging(data, x, y)
}
\arguments{
\item{data}{Dataframe containing the variables in the model.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}
}
\value{
List containing training dataframe and list with binary respose as 0 if the observations have been selected for training and 0 in any other case.
}
\description{
Bootstrap aggregating for data.
}
