% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{barplot_importance}
\alias{barplot_importance}
\title{Barplot variable importance}
\usage{
barplot_importance(m)
}
\arguments{
\item{m}{Dataframe with the importance of each variable.}
}
\value{
Barplot that represents each variable in the x-axis and its importance in the y-axis. It returns the numeric values too.
}
\description{
This function generates a barplot with the importance of each variable.
}
