% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bests.R
\name{bestEAT}
\alias{bestEAT}
\title{Tuning an EAT model}
\usage{
bestEAT(
  training,
  test,
  x,
  y,
  numStop = NULL,
  fold = NULL,
  max.depth = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{training}{Training dataframe or matrix containing the variables in the model for model construction.}

\item{test}{Test dataframe or matrix containing the variables in the model for model assessment.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{numStop}{Vector. Set of minimun number of observations in a node for a split to be attempted.}

\item{fold}{Vector. Set of number of folds in which is divided the dataset to apply cross-validation during the pruning.}

\item{max.depth}{Vector. Integer. Maximum number of leaf nodes.}

\item{na.rm}{Logical. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
Dataframe in which each row corresponds to a given set of hyperparameters with its corresponding root mean square error.
}
\description{
This funcion calculates the mean square error for a Efficiency Analysis Tree built with a set of given hyperparameters.
}
\examples{

data("PISAindex")

n <- nrow(PISAindex) # Observations in the dataset
t_index <- sample(1:n, n * 0.7) # Training indexes
training <- PISAindex[t_index, ] # Training set
test <- PISAindex[-t_index, ] # Test set

bestEAT(training = training, 
        test = test,
        x = 6:9,
        y = 3,
        numStop = c(3, 5, 7, 10),
        fold = c(5, 7, 10))

}
