% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bests.R
\name{bestRFEAT}
\alias{bestRFEAT}
\title{Tuning an RFEAT model}
\usage{
bestRFEAT(training, test, x, y, numStop, m, s_mtry, na.rm = TRUE)
}
\arguments{
\item{training}{Training dataframe or matrix containing the variables in the model for model construction.}

\item{test}{Test dataframe or matrix containing the variables in the model for model assessment.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{numStop}{Vector. Set of minimun number of observations in a node for a split to be attempted.}

\item{m}{Vector. Set of number of trees to be build.}

\item{na.rm}{Logical. If \code{TRUE}, \code{NA} rows are omitted.}

\item{s_mtry.}{Character vector. Set of options for selecting number of inputs to be selected in each split.}
}
\value{
Dataframe in which each row corresponds to a given set of hyperparameters with its corresponding mean square error.
}
\description{
This funcion calculates the mean square error for a Random Forest of Efficiency Analysis Tree built with a set of given hyperparameters.
}
\examples{

data("PISAindex")

n <- nrow(PISAindex) # Observations in the dataset
t_index <- sample(1:n, n * 0.7) # Training indexes
training <- PISAindex[t_index, ] # Training set
test <- PISAindex[-t_index, ] # Test set

bestRFEAT(training = training, 
          test = test,
          x = 6:9,
          y = 3,
          numStop = c(3, 5, 7, 10),
          m = c(20, 40, 60),
          s_mtry = c("Breiman", "1", "2"))

}
