% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestModels.R
\name{bestRFEAT}
\alias{bestRFEAT}
\title{Tuning a Random Forest + Efficiency Analysis Trees model}
\usage{
bestRFEAT(
  training,
  test,
  x,
  y,
  numStop = 5,
  m = 50,
  s_mtry = c("5", "BRM"),
  na.rm = TRUE
)
}
\arguments{
\item{training}{Training dataframe or matrix containing the variables for model construction.}

\item{test}{Test dataframe or matrix containing the variables for model assessment.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{numStop}{Vector. Set of minimun number of observations in a node for a split to be attempted.}

\item{m}{Vector. Set of number of trees to be build.}

\item{s_mtry}{Character vector. Set of options for selecting the number of inputs to be selected in each split.}

\item{na.rm}{Logical. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
Dataframe in which each row corresponds to a given set of hyperparameters and the root mean square error (RMSE).
}
\description{
This funcion calculates the root mean square error (RMSE) for a Random Forest + Efficiency Analysis Tree model built with a set of given hyperparameters.
}
\examples{

data("PISAindex")

n <- nrow(PISAindex) # Observations in the dataset
t_index <- sample(1:n, n * 0.7) # Training indexes
training <- PISAindex[t_index, ] # Training set
test <- PISAindex[- t_index, ] # Test set

bestRFEAT(training = training, 
          test = test,
          x = 6:9,
          y = 3,
          numStop = c(3, 5, 7, 10),
          m = c(20, 40, 60),
          s_mtry = c("BRM", "1", "2"))

}
