% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencyCEAT.R
\name{CEAT_WAM}
\alias{CEAT_WAM}
\title{Weighted Additive Model for a convex Efficiency Analysis Tree model}
\usage{
CEAT_WAM(j, scores, x_k, y_k, atreeTk, ytreeTk, nX, nY, N_leaves, weights)
}
\arguments{
\item{j}{Integer. Number of DMUs.}

\item{scores}{Matrix. Empty matrix for scores.}

\item{x_k}{Dataframe. Set of input variables.}

\item{y_k}{Dataframe. Set of output variables.}

\item{atreeTk}{Matrix. Set of "a" Pareto-coordinates.}

\item{ytreeTk}{Matrix. Set of predictions.}

\item{nX}{Integer. Number of inputs.}

\item{nY}{Integer. Number of outputs.}

\item{N_leaves}{Integer. Number of leaf nodes.}

\item{weights}{Character. \code{"MIP"} for Measure of Inefficiency Proportion or \code{"RAM"} for Range Adjusted Measure of Inefficiency.}
}
\value{
A numerical vector with scores.
}
\description{
Weighted Additive Model for a convex Efficiency Analysis Tree model.
}
