% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEAT.R
\name{check}
\alias{check}
\title{Check Efficiency Analysis Trees.}
\usage{
check(tree)
}
\arguments{
\item{tree}{A list containing EAT nodes.}
}
\value{
Message indicating if the tree is acceptable or warning in case of breaking any Pareto-dominance relationship.
}
\description{
This function verifies if a specific tree keeps to Pareto-dominance properties.
}
