% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_program.R
\name{create_fit_func}
\alias{create_fit_func}
\title{Create a fitness function to estimate a program (R function)}
\usage{
create_fit_func(loss_function, X, y, parallel)
}
\arguments{
\item{loss_function}{function resulting from a call to 
\code{\link{create_loss_func}}.}

\item{X}{model.matrix}

\item{y}{outcome}

\item{parallel}{Optional Logical vector length one. Default is \code{parallel 
= FALSE}; \code{parallel = TRUE} can be slower if the data set is small 
relative to the numner of population evolutions desired}
}
\value{
Returns a function that takes in a function as an argument and then 
  assigns it a numeric fitness value we want to minimize. Use this in
  \code{\link{estimate_program}}.
}
\description{
Create a fitness function to estimate a program (R function)
}

