% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_program.R
\name{create_func_set}
\alias{create_func_set}
\title{Create function set}
\usage{
create_func_set(func_list, link)
}
\arguments{
\item{func_list}{List where each element is a length one character vector.}

\item{link}{Length one character vector to pass to create_link_func}
}
\value{
Returns a function set.
}
\description{
Create function set
}

