% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{deepEAT}
\alias{deepEAT}
\title{Deep Efficiency Analysis Trees}
\usage{
deepEAT(data, x, y, numStop = 5, max.depth = NULL, max.leaves = NULL)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{numStop}{Integer. Minimum number of observations in a node for a split to be attempted.}

\item{max.depth}{Integer. Depth of the tree.}

\item{max.leaves}{Integer. Maximum number of leaf nodes.}
}
\value{
List containing each possible pruning for the deep tree and its associated alpha value.
}
\description{
This function creates a deep Efficiency Analysis Tree and a set of possible prunings by the weakest-link pruning procedure.
}
