% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{deepEAT}
\alias{deepEAT}
\title{Deep Efficiency Analysis Trees}
\usage{
deepEAT(data, x, y, numStop = 5, max.depth = NULL, max.leaves = NULL)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{numStop}{Integer. Minimun number of observations in a node for a split to be attempted.}

\item{max.depth}{Integer. Maximum number of leaf nodes.}
}
\value{
List containing each possible pruning for the deep tree and its alpha associated.
}
\description{
This function creates a deep tree and a set of possible prunings by weakest-link pruning procedure.
}
