% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{deepEAT}
\alias{deepEAT}
\title{Deep Efficiency Analysis Tree}
\usage{
deepEAT(data, x, y, numStop)
}
\arguments{
\item{data}{Data to be used.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{numStop}{Minimun number of observations on a node to be splitted.}
}
\value{
List with each possible pruning for the deep tree and its alpha associated.
}
\description{
This function creates a deep tree and a set of possible prunings.
}
