% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eat.R
\docType{package}
\name{eat}
\alias{eat}
\alias{eat-package}
\title{eat: An R package for data-driven agent-based modeling}
\description{
\code{eat} functions:
\enumerate{
\item \code{\link{create_sample}}
\item \code{\link{sobol_sa}}
\item \code{\link{plot_sobol_fo}}
\item \code{\link{plot_sobol_total}}
}
}

