% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{EAT}
\alias{EAT}
\title{Efficiency Analysis Tree}
\usage{
EAT(data, x, y, fold = 5, numStop = 5)
}
\arguments{
\item{data}{Data to be used.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{fold}{Parts in which is divided the original data set to do Cross-Validation.}

\item{numStop}{Minimun number of observations on a node to be splitted.}
}
\value{
Best pruned tree by weakest-link pruning.
}
\description{
This function prunes the tree-object of deepEAT function by weakest-link pruning method.
}
