% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{EAT}
\alias{EAT}
\title{Efficiency Analysis Trees}
\usage{
EAT(
  data,
  x,
  y,
  numStop = 5,
  fold = 5,
  max.depth = NULL,
  max.leaves = NULL,
  na.rm = T
)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{numStop}{Integer. Minimun number of observations in a node for a split to be attempted.}

\item{fold}{Integer. Number of folds in which is divided the dataset to apply cross-validation during the pruning.}

\item{max.depth}{Integer. Depth of the tree.}

\item{max.leaves}{Integer. Maximum number of leaf nodes.}

\item{na.rm}{Logical. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
An EAT object.
}
\description{
This function estimates a stepped production frontier through regression trees.
}
\details{
The EAT function generates a regression tree model based on CART \insertCite{breiman1984}{eat} under a new approach that guarantees the obtaining of a stepped production frontier that fulfills the property of free disposability. This frontier shares the aforementioned aspects with the FDH frontier \insertCite{deprins1984}{eat} but enhances some of its disadvantages such as the overfitting problem or the underestimation of technical inefficiency. More details in \insertCite{esteve2020;textual}{eat}.
}
\examples{
# ====================== #
# Single output scenario #
# ====================== #

simulated <- eat:::Y1.sim(N = 50, nX = 3)
EAT(data = simulated, x = c(1, 2, 3), y = 4, numStop = 10, fold = 5)

# ====================== #
#  Multi output scenario #
# ====================== #

simulated <- eat:::X2Y2.sim(N = 50, border = 0.1)
EAT(data = simulated, x = c(1,2), y = c(3, 4), numStop = 10, fold = 7)

}
\references{
\insertRef{breiman1984}{eat} \cr
\cr
\insertRef{deprins1984}{eat} \cr
\cr
\insertRef{esteve2020}{eat}
}
