% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{EAT_object}
\alias{EAT_object}
\title{EAT object}
\usage{
EAT_object(
  data,
  x,
  y,
  register_names,
  numStop,
  fold,
  max.depth,
  max.leaves,
  na.rm,
  tree
)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{register_names}{String vector. Data rownames.}

\item{numStop}{Integer. Minimum number of observations in a node for a split to be attempted.}

\item{fold}{Integer. Set of number of folds in which the dataset to apply cross-validation during the pruning is divided.}

\item{max.depth}{Integer. Maximum number of leaf nodes.}

\item{max.leaves}{Integer. Depth of the tree.}

\item{na.rm}{Logical. If True, NA rows are omitted. If False, an error occurs in case of NA rows.}

\item{tree}{A list containing the nodes of the EAT pruned tree.}
}
\value{
An EAT object
}
\description{
This function saves information about the EAT model
}
