% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT_plot.R
\name{EAT_plot}
\alias{EAT_plot}
\title{Efficiency Analysis Trees Plot}
\usage{
EAT_plot(tree, data, x, y)
}
\arguments{
\item{tree}{Tree-structure object of EAT function.}

\item{data}{Data to be used. Dataframe or matrix objects are acceptable.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indxes in data.}
}
\value{
Plot object with the tree-structure.
}
\description{
This function creates a plot with the tree-structure object result of EAT function
}
