% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT_plot.R
\name{EAT_plot}
\alias{EAT_plot}
\title{Efficiency Analysis Trees Plot}
\usage{
EAT_plot(object)
}
\arguments{
\item{object}{An EAT object.}
}
\value{
Plot object with the following elements for each node:
id: node index \cr
\cr
R: mean square error \cr
\cr
Samples: number of observations in a node \cr
\cr
Variable: splitting variable \cr
\cr
y: output prediction
}
\description{
Plot an EAT model.
}
