% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency_barplot}
\alias{efficiency_barplot}
\title{Efficiency Scores Bar Plot}
\usage{
efficiency_barplot(object, scores, color = T)
}
\arguments{
\item{object}{An EAT object.}

\item{scores}{Dataframe with scores.}

\item{color}{Logical. If True, observations with same efficient level output are painted in the same color.}
}
\value{
Bar plot with efficiency scores
}
\description{
This function returns a bar plot with scores
}
