% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency_density}
\alias{efficiency_density}
\title{Efficiency Scores Density Plot}
\usage{
efficiency_density(object, scores, FDH = TRUE)
}
\arguments{
\item{object}{An EAT object.}

\item{scores}{Dataframe with scores.}

\item{FDH}{Logical. If FDH = TRUE, density plot for FDH is showed too.}
}
\value{
Density plot for scores
}
\description{
Density plot for scores. EAT scores in red and FDH scores in blue.
}
