% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_plots.R
\name{efficiency_density}
\alias{efficiency_density}
\title{Efficiency Scores Density Plot}
\usage{
efficiency_density(scores_EAT, scores_FDH = NULL)
}
\arguments{
\item{scores_EAT}{Dataframe with EAT scores.}

\item{scores_FDH}{Optional. Dataframe with FDH scores.}
}
\value{
Density plot for scores and data.frame with numeric results.
}
\description{
Density plot for scores. EAT scores in red and FDH scores in blue.
}
