% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEAT.R
\name{efficiency_RFEAT}
\alias{efficiency_RFEAT}
\title{RFEAT Efficiency Scores}
\usage{
efficiency_RFEAT(data, x, y, object)
}
\arguments{
\item{data}{Dataframe for which the efficiency score is calculated.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{object}{A RFEAT object}
}
\value{
Number of rows, matrix for scoring, matrix of inputs, matrix of outputs, a Pareto-coordinates, predictions, number of inputs, number of outputs and number of leaf nodes.
}
\description{
This function calculates the efficiency scores for each DMU by an RFEAT model.
}
