% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEAT.R
\name{efficiency_RFEAT}
\alias{efficiency_RFEAT}
\title{Efficiency Objects of Random Forest EAT}
\usage{
efficiency_RFEAT(data, forest_err, y)
}
\arguments{
\item{data}{Data to be used.}

\item{forest_err}{List was returned by RFEAT function}

\item{y}{Column output indexes in data.}
}
\value{
Number of rows, matrix for scoring, matrix of inputs, matrix of outputs, a Pareto-coordinates, predictions, number of inputs, number of outputs and number of leaf nodes.
}
\description{
Efficiency score of Random Forest EAT
}
