% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_plots.R
\name{efficiencyDensity}
\alias{efficiencyDensity}
\title{Efficiency Scores Density Plot}
\usage{
efficiencyDensity(scores_EAT, scores_FDH = NULL)
}
\arguments{
\item{scores_EAT}{Dataframe with EAT scores.}

\item{scores_FDH}{Optional. Dataframe with FDH scores.}
}
\value{
Density plot for scores and data.frame with numeric results.
}
\description{
Density plot for scores. EAT scores in red and FDH scores in blue.
}
\examples{

X1 <- runif(50, 1, 10)
X2 <- runif(50, 2, 10)
Y1 <- log(X1) + 3 - abs(rnorm(50, mean = 0, sd = 0.4))
Y2 <- log(X1) + 2 - abs(rnorm(50, mean = 0, sd = 0.7))

simulated <- data.frame(x1 = X1, x2 = X2, y1 = Y1, y2 = Y2)

EAT_model <- EAT(data = simulated, x = c(1,2), y = c(3, 4), numStop = 5, fold = 5)

EAT_scores <- efficiencyEAT(data = simulated, x = c(1, 2), y = c(3, 4), object = EAT_model, 
                            scores_model = "EAT_BCC_out", r = 2, na.rm = TRUE)

FDH_scores <- efficiencyFDH(data = simulated, x = c(1, 2), y = c(3, 4), 
                            scores_model = "FDH_BCC_out", r = 2, na.rm = TRUE)

efficiencyDensity(scores_EAT = EAT_scores$EAT_BCC_out, scores_FDH = FDH_scores$FDH_BCC_out)

}
