% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiencyEAT.R
\name{efficiencyEAT}
\alias{efficiencyEAT}
\title{Efficiency Analysis Tree Efficiency Scores}
\usage{
efficiencyEAT(
  data,
  x,
  y,
  object,
  scores_model,
  r = 2,
  FDH = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{Dataframe for which the efficiency score is calculated.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{object}{An EAT object.}

\item{scores_model}{Mathematic programming model to calculate scores. 
\itemize{
\item{\code{BCC_out}} BCC model. Output-oriented.
\item{\code{BCC_in}}  BCC model. Input-oriented.
\item{\code{DDF}}     Directional Distance Function.
\item{\code{RSL_out}} Rusell model. Output-oriented.
\item{\code{RSL_in}}  Rusell model. Input-oriented.
\item{\code{WAM}}     Weighted Additive Model.
}}

\item{r}{Integer. Decimal units for scores.}

\item{FDH}{Logical. If \code{TRUE}, FDH scores are calculated with the programming model selected in \code{scores_model}.}

\item{na.rm}{Logical. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
Dataframe with input variables and efficiency scores by an EAT model.
}
\description{
This function calculates the efficiency scores for each DMU by an Efficiency Analysis Tree model.
}
\examples{

simulated <- eat:::X2Y2.sim(N = 50, border = 0.2)
EAT_model <- EAT(data = simulated, x = c(1,2), y = c(3, 4))

efficiencyEAT(data = simulated, x = c(1, 2), y = c(3, 4), object = EAT_model, 
              scores_model = "BCC_out", r = 2, FDH = TRUE, na.rm = TRUE)

}
