% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_plots.R
\name{efficiencyJitter}
\alias{efficiencyJitter}
\title{Efficiency Scores Jitter Plot}
\usage{
efficiencyJitter(object, scores_EAT, scores_model, upb = NULL, lwb = NULL)
}
\arguments{
\item{object}{An EAT object.}

\item{scores_EAT}{Dataframe with scores.}

\item{scores_model}{Mathematic programming model of scores_EAT. 
\itemize{
\item{\code{EAT_BCC_out}} BBC model. Output orientation.
\item{\code{EAT_BCC_in}}  BBC model. Input orientation.
\item{\code{EAT_DDF}}     Directional distance model.
\item{\code{EAT_RSL_out}} Rusell model. Output orientation
\item{\code{EAT_RSL_in}}  Rusell model. Input orientation.
\item{\code{EAT_WAM}}     Weighted Additive model.
}}

\item{upb}{Numeric. Upper bound for labeling.}

\item{lwb}{Numeric. Lower bound for labeling.}
}
\value{
Jitter plot with DMUs and scores.
}
\description{
This function returns a jitter plot. DMUs are grouped in nodes (same input paradigm) and its scores are showed. A black point represents the mean score in a group and the line the standard deviation. Finally, efficient DMUs are labeled and the user can specify an upper bound (upb) and a lower bound (lwb) in order to show extra labels.
}
\examples{

X1 <- runif(50, 1, 10)
X2 <- runif(50, 2, 10)
Y1 <- log(X1) + 3 - abs(rnorm(50, mean = 0, sd = 0.4))
Y2 <- log(X1) + 2 - abs(rnorm(50, mean = 0, sd = 0.7))

simulated <- data.frame(x1 = X1, x2 = X2, y1 = Y1, y2 = Y2)

EAT_model <- EAT(data = simulated, x = c(1,2), y = c(3, 4), numStop = 5, fold = 5)

EAT_scores <- efficiencyEAT(data = simulated, x = c(1, 2), y = c(3, 4), object = EAT_model,
                            scores_model = "EAT_BCC_out", r = 2, na.rm = TRUE)

efficiencyJitter(object = EAT_model, scores_EAT = EAT_scores, scores_model = "EAT_BCC_out")

}
