% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_plots.R
\name{efficiencyJitter}
\alias{efficiencyJitter}
\title{Efficiency Scores Jitter Plot}
\usage{
efficiencyJitter(object, scores_EAT, scores_model, upb = NULL, lwb = NULL)
}
\arguments{
\item{object}{An EAT object.}

\item{scores_EAT}{Dataframe with scores.}

\item{scores_model}{Mathematic programming model of scores_EAT. 
\itemize{
\item{\code{BCC_out}} BBC model. Output orientation.
\item{\code{BCC_in}}  BBC model. Input orientation.
\item{\code{DDF}}     Directional distance model.
\item{\code{RSL_out}} Rusell model. Output orientation
\item{\code{RSL_in}}  Rusell model. Input orientation.
\item{\code{WAM}}     Weighted Additive model.
}}

\item{upb}{Numeric. Upper bound for labeling.}

\item{lwb}{Numeric. Lower bound for labeling.}
}
\value{
Jitter plot with DMUs and scores.
}
\description{
This function returns a jitter plot. DMUs are grouped in nodes (same input paradigm) and its scores are showed. A black point represents the mean score in a group and the line the standard deviation. Finally, efficient DMUs are labeled and the user can specify an upper bound (upb) and a lower bound (lwb) in order to show extra labels.
}
\examples{

simulated <- eat:::X2Y2.sim(N = 50, border = 0.2)
EAT_model <- EAT(data = simulated, x = c(1,2), y = c(3, 4))

EAT_scores <- efficiencyEAT(data = simulated, x = c(1, 2), y = c(3, 4), object = EAT_model,
                            scores_model = "BCC_out", r = 2, na.rm = TRUE)

efficiencyJitter(object = EAT_model, scores_EAT = EAT_scores, scores_model = "BCC_out")

}
