% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFEAT.R
\name{efficiencyRFEAT}
\alias{efficiencyRFEAT}
\title{RFEAT Efficiency Scores}
\usage{
efficiencyRFEAT(data, x, y, object, r = 2, FDH = TRUE, na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe for which the efficiency score is calculated.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{object}{A RFEAT object.}

\item{r}{Integer. Decimal units for scores.}

\item{na.rm}{Logical. If \code{TRUE}, \code{NA} rows are omitted.}

\item{FDH.}{Logical. If \code{TRUE}, FDH scores are calculated with the programming model selected in \code{scores_model}.}
}
\value{
Dataframe with input variables and efficiency scores by a RFEAT model.
}
\description{
This function calculates the efficiency scores for each DMU by an RFEAT model.
}
\examples{

X1 <- runif(50, 1, 10)
X2 <- runif(50, 2, 10)
Y1 <- log(X1) + 3 - abs(rnorm(50, mean = 0, sd = 0.4))
Y2 <- log(X1) + 2 - abs(rnorm(50, mean = 0, sd = 0.7))

simulated <- data.frame(x1 = X1, x2 = X2, y1 = Y1, y2 = Y2)

RFEAT_model <- RFEAT(data = simulated, x = c(1,2), y = c(3, 4), numStop = 5,
                     m = 50, s_mtry = "Breiman", na.rm = TRUE)

efficiencyRFEAT(data = simulated, x = c(1, 2), y = c(3, 4), object = RFEAT_model)

}
