% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_program.R
\name{estimate_program}
\alias{estimate_program}
\title{Estimate a program (R function) from data.}
\usage{
estimate_program(data, loss = c("log_lik", "rmse", "identity",
  "identity_multi_class"), link = c("logit", "probit", "cauchit", "identity"),
  mins = 2, parallel = FALSE, cores = NULL)
}
\arguments{
\item{data}{data.frame where (i.) the last column is named \code{"outcome"}
and has the outcome variable we are evolving programs to predict; (ii.) all
other columns (there must be at least one other column) that precede the
\code{"outcome"} column are named columns containing the predictor
variables, which can be of any type. The order matters: when using the
evolved program for predictions the (named) arguments to the function will
be the predictor variables in the order they are supplied to this
\code{estimate_program()} function.}

\item{loss}{Optional Character vector length one}

\item{link}{Optional Character vector length one}

\item{mins}{Optional Integer vector length one}

\item{parallel}{Optional Logical vector length one. Default is \code{parallel
= FALSE}; \code{parallel = TRUE} can be slower if the data set is small
relative to the numner of population evolutions desired}

\item{cores}{Optional Integer vector length one}
}
\value{
The function returns an S4 object. See \linkS4class{estimate_program}
 for the details of the \code{slots} (objects) that this type of object will
 have.
}
\description{
\code{estimate_program} uses \code{rgp}'s \code{rgp::typedGeneticProgramming}.
}

