% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{FDH}
\alias{FDH}
\title{Free Disposal Hull}
\usage{
FDH(data, xarray, y)
}
\arguments{
\item{data}{Data frame or matrix.}

\item{xarray}{Input columns}

\item{yarray}{Output columns}
}
\description{
...
}
