% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontier.R
\name{frontier}
\alias{frontier}
\title{EAT frontier graph for a single input and output scenario}
\usage{
frontier(
  object,
  train.data = FALSE,
  train.color = "black",
  pch = 19,
  size = 1,
  rwn = FALSE
)
}
\arguments{
\item{object}{An EAT object.}

\item{train.data}{Logical. If \code{TRUE}, training DMUs are displayed.}

\item{train.color}{String. Color for training DMUs.}

\item{pch}{Integer. Point shape.}

\item{size}{Integer. Point size.}

\item{rwn}{Logical. If \code{TRUE}, rownames are displayed instead of points.}
}
\value{
Plot with estimated production frontier
}
\description{
This function displays a plot with the frontier estimated by EAT for a scenario corresponding to one input and one output.
}
\examples{

data("PISAindex")
n <- nrow(PISAindex)
t_index <- sample(1:n, n * 0.8)
training <- PISAindex[t_index, ]
test <- PISAindex[-t_index, ]

single_model <- EAT(data = training, x = 6, y = 3)

frontier <- frontier(object = single_model, train.data = TRUE, train.color = "#F8766D",
                     rwn = TRUE, size = 1.5)

plot(frontier)

}
