% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontier.R
\name{frontier}
\alias{frontier}
\title{EAT frontier graph for a single input and output scenario}
\usage{
frontier(
  object,
  train.data = FALSE,
  train.color = "black",
  pch = 19,
  size = 1,
  rwn = FALSE
)
}
\arguments{
\item{object}{An EAT object.}

\item{train.data}{Logical. If \code{TRUE}, training DMUs are displayed.}

\item{train.color}{String. Color for training DMUs.}

\item{pch}{Integer. Point shape.}

\item{size}{Integer. Point size.}

\item{rwn}{Logical. If \code{TRUE}, rownames are displayed instead of points.}
}
\value{
Estimated production frontier
}
\description{
This function creates a plot with the frontier estimated by EAT for the scenario corresponding to one input and one output.
}
