% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontier.R
\name{frontier}
\alias{frontier}
\title{EAT frontier graph for a single input and output scenario}
\usage{
frontier(
  object,
  FDH = FALSE,
  train.data = FALSE,
  train.color = "black",
  pch = 19,
  size = 1,
  rwn = FALSE
)
}
\arguments{
\item{object}{An EAT object.}

\item{FDH}{Logical. If \code{TRUE}, FDH frontier is displayed.}

\item{train.data}{Logical. If \code{TRUE}, training DMUs are displayed.}

\item{train.color}{String. Color for training DMUs.}

\item{pch}{Integer. Point shape.}

\item{size}{Integer. Point size.}

\item{rwn}{Logical. If \code{TRUE}, rownames are displayed.}
}
\value{
Plot with estimated production frontier
}
\description{
This function displays a plot with the frontier estimated by Efficiency Analysis Trees in a scenario of one input and one output.
}
\examples{

simulated <- eat:::Y1.sim(N = 50, nX = 1)

model <- EAT(data = simulated,
             x = 1,
             y = 2)

frontier <- frontier(object = model,
                     FDH = TRUE, 
                     train.data = TRUE,
                     rwn = TRUE)
plot(frontier)

}
