% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontier.R
\name{frontier}
\alias{frontier}
\title{Single input and output scenario EAT frontier}
\usage{
frontier(
  t,
  train.data = FALSE,
  train.color = "black",
  pch = 19,
  rownames = FALSE,
  rwn.size = 3
)
}
\arguments{
\item{t}{An EAT object.}

\item{train.data}{Logical. If T, training DMUs are displayed.}

\item{train.color}{String. Color for training DMUs.}

\item{pch}{Integer. Point shape.}

\item{rownames}{Logical. If T, row names are displayed instead of points.}

\item{rwn.size}{Integer. Rowname size}
}
\value{
Scatter plot with DMUs and the frontier predicted by EAT.
}
\description{
This function creates a plot with the frontier estimated by EAT for the scenario corresponding to one input and one output.
}
