% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingEAT.R
\name{imp_var_EAT}
\alias{imp_var_EAT}
\title{Breiman's Variable Importance}
\usage{
imp_var_EAT(data, tree, x, y, digits)
}
\arguments{
\item{data}{Data from EAT object.}

\item{tree}{Tree from EAT object.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{digits}{Decimal units.}
}
\value{
A dataframe with the best split for each node and its variable importance.
}
\description{
This function recalculates all the possible splits, with the exception of the one being used, and for each node and variable gets the best split based on their degree of importance.
}
