% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFEAT.R
\name{imp_var_RFEAT}
\alias{imp_var_RFEAT}
\title{Importance variable of xj in Random Forest EAT}
\usage{
imp_var_RFEAT(object, r = 2)
}
\arguments{
\item{object}{An RFEAT object}

\item{r}{Integer. Decimal units.}
}
\value{
List of importance of inputs xj
}
\description{
Importance variable of xj in Random Forest EAT
}
