% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingRFEAT.R
\name{imp_var_RFEAT}
\alias{imp_var_RFEAT}
\title{Variable Importance through Random Forest + Efficiency Analysis Trees}
\usage{
imp_var_RFEAT(object, digits = 2)
}
\arguments{
\item{object}{A RFEAT object}

\item{digits}{Integer. Decimal units.}
}
\value{
Vector of input importance scores
}
\description{
Variable Importance through Random Forest + Efficiency Analysis Trees.
}
