% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitEAT.R
\name{isFinalNode}
\alias{isFinalNode}
\title{Is Final Node}
\usage{
isFinalNode(obs, data, numStop)
}
\arguments{
\item{obs}{Observation in the evaluated node.}

\item{data}{Data with predictive variable.}

\item{numStop}{Minimum number of observations in a node to be split.}
}
\value{
True if the node is a final node and false in any other case.
}
\description{
This function evaluates a node and checks if it fulfills the conditions to be a final node.
}
