% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEAT.R
\name{layout}
\alias{layout}
\title{Layout for nodes in plotEAT}
\usage{
layout(py)
}
\arguments{
\item{py}{a party object.}
}
\value{
Dataframe with suitable modifications of the nodes layout.
}
\description{
This function modifies the coordinates of the nodes in the plotEAT function to overcome overlapping.
}
