% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT_plot.R
\name{layout}
\alias{layout}
\title{Layout for nodes in EAT_plot}
\usage{
layout(py)
}
\arguments{
\item{py}{party object.}
}
\value{
Suitable modification of layout dataframe.
}
\description{
This function modifies the coordinates of the nodes in the EAT_plot function to overcome overlapping.
}
