% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{M_Breiman}
\alias{M_Breiman}
\title{Importance of Breiman}
\usage{
M_Breiman(data, tree, x, y, r = 2)
}
\arguments{
\item{data}{Data to be used.}

\item{tree}{Tree-object of EAT function.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{r}{Decimal units.}
}
\value{
Dataframe with one column and the importance of each variable in rows.
}
\description{
This function evaluates the importance of each predictor variable by Breiman method.
}
