% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def_estimate_program.R
\docType{class}
\name{model_program-class}
\alias{model_program-class}
\title{An S4 class to return the best results of using the estimate_program 
function.}
\description{
An S4 class to return the best results of using the estimate_program 
function.
}
\section{Slots}{

\describe{
\item{\code{func}}{function}

\item{\code{levels}}{Numeric vector length one.}

\item{\code{Terms}}{terms used for fitting the model, got these from \code{model.frame}.}
}}

