% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitEAT.R
\name{mse}
\alias{mse}
\title{Mean Square Error}
\usage{
mse(data, t, y)
}
\arguments{
\item{data}{Data to be used.}

\item{t}{A given node.}

\item{y}{Column output indexes in data.}
}
\value{
Mean Square Error on a node.
}
\description{
This function calculates the Mean Square Error between the predicted value and the observations in a given node.
}
