% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_def_cv_abm.R
\name{performance}
\alias{performance}
\title{Turns modeling related object into useful outputs for summary or plotting.}
\usage{
performance(results, output = character(1))
}
\arguments{
\item{results}{Modeling related object, e.g. S4 \code{cv_abm} object.}

\item{output}{Optional character vector length one indicating what the 
desired output is, e.g. \code{c("MSE", "cor", "cor_pval", "SE", 
"plot")}.}
}
\description{
Turns modeling related object into useful outputs for summary or plotting.
}

