% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{performance}
\alias{performance}
\title{Performance of Efficiency Analysis Trees model}
\usage{
performance(object)
}
\arguments{
\item{object}{An EAT object.}
}
\value{
List with the root mean square error (RMSE), the mean absolute error (MAE) and the bias (Bias). Additionally, for each leaf node, centralization and dispersion measures and the RMSE are computed.
}
\description{
This function returns a set of common metrics for regression trees for the EAT model as a whole and disaggregated for the leaf nodes.
}
\examples{
simulated <- eat:::Y1.sim(N = 50, nX = 3)
model <- EAT(data = simulated, x = c(1, 2, 3), y = 4, numStop = 10, fold = 5)
performance(model)

}
