% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictorEAT.R
\name{posIdNode}
\alias{posIdNode}
\title{Position of the node}
\usage{
posIdNode(tree, idNode)
}
\arguments{
\item{tree}{A list containing EAT nodes.}

\item{idNode}{Id of a specific node.}
}
\value{
Position of the node or -1 if it is not found.
}
\description{
This function finds the node where a register is located.
}
