% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predict}
\alias{predict}
\title{Model prediction}
\usage{
predict(t, newdata)
}
\arguments{
\item{t}{An EAT object.}

\item{newdata}{Dataframe. Set of input variables to predict on.}
}
\value{
Data frame with predicted values with as many columns as outputs.
}
\description{
This function predicts the expected output by an EAT object.
}
