% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predict_EAT}
\alias{predict_EAT}
\title{Model prediction for EAT}
\usage{
predict_EAT(object, newdata)
}
\arguments{
\item{object}{An EAT object.}

\item{newdata}{Dataframe. Set of input variables to predict on.}
}
\value{
Data frame with the original data and the predicted values.
}
\description{
This function predicts the expected output by an EAT object.
}
