% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDH.R
\name{predictFDH}
\alias{predictFDH}
\title{Model prediction for Free Disposal Hull}
\usage{
predictFDH(data, x, y)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}
}
\value{
Data frame with the original data and the predicted values through a Free Disposal Hull model.
}
\description{
This function predicts the expected output by a Free Disposal Hull model.
}
