% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{preProcess}
\alias{preProcess}
\title{Data Preprocessing for Efficiency Analysis Trees}
\usage{
preProcess(
  data,
  x,
  y,
  numStop = 5,
  fold = 5,
  max.depth = NULL,
  max.leaves = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{data}{\code{data.frame} or \code{matrix} containing the variables in the model.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{numStop}{Minimum number of observations in a node for a split to be attempted.}

\item{fold}{Set of number of folds in which the dataset to apply cross-validation during the pruning is divided.}

\item{max.depth}{Depth of the tree.}

\item{max.leaves}{Maximum number of leaf nodes.}

\item{na.rm}{\code{logical}. If \code{TRUE}, \code{NA} rows are omitted.}
}
\value{
It returns a \code{data.frame} in the required format.
}
\description{
This function arranges the data in the required format and displays error messages.
}
