% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preProcess}
\alias{preProcess}
\title{Data preprocessing}
\usage{
preProcess(data, x, y, na.rm = T)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{na.rm}{Logical. If True, NA rows are omitted. If False, an error occurs in case of NA rows.}
}
\value{
Data processed in the [X, Y] format with only allowed classes.
}
\description{
This function displays error messages in relation with data such as presence of NA values or not allowed classes of data.
}
