% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{print_results}
\alias{print_results}
\title{Print results}
\usage{
print_results(tree, data, x)
}
\arguments{
\item{tree}{Tree structure.}

\item{data}{Data to be used.}

\item{x}{Column input indexes in data.}
}
\value{
Output print
}
\description{
This function prints for each terminal node: the number of observations in it, the efficiency level of the outputs and the mean square error.
}
