% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{RBranch}
\alias{RBranch}
\title{Branch Pruning}
\usage{
RBranch(t, tree)
}
\arguments{
\item{t}{List. A given EAT node.}

\item{tree}{A list containing EAT nodes.}
}
\value{
List containing the sum of the errors of the child nodes of the pruned node and the total number of leaf nodes that come from it.
}
\description{
This function computes the error of a branch as the sum of the errors of its child nodes.
}
