% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{RBranch}
\alias{RBranch}
\title{Branch Pruning}
\usage{
RBranch(t, tree)
}
\arguments{
\item{t}{\code{list}. A given EAT node.}

\item{tree}{A \code{list} containing the EAT nodes.}
}
\value{
A \code{list} containing (1) the sum of the errors of the child nodes of the pruned node and (2) the total number of leaf nodes that come from it.
}
\description{
This function computes the error of a branch as the sum of the errors of its child nodes.
}
