% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{RBranch}
\alias{RBranch}
\title{RBranch}
\usage{
RBranch(t, tree)
}
\arguments{
\item{t}{A given node.}

\item{tree}{List structure with the tree nodes.}
}
\value{
List structure with the sum of the errors of the children of the given node and the total number of leaf nodes that comes from it.
}
\description{
Recursive function that calculates the error of a node as the sum of the errors of its children and the total number of leaf nodes that come from it.
}
