% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{RCV}
\alias{RCV}
\title{RCV}
\usage{
RCV(N, Lv, y, alphaIprim, fold, TAiv)
}
\arguments{
\item{N}{Number of rows in data.}

\item{Lv}{Test set.}

\item{y}{Column output indexes in data.}

\item{alphaIprim}{Alpha obtained as the square root of the product of two consecutives alpha values in tree_alpha list. It is used to find the best pruing tree.}

\item{fold}{Parts in which is divided the original data set to do Cross-Validation.}

\item{TAiv}{List with each possible pruning for the deep tree generated with train set and its alpha values associated.}
}
\value{
Set of best prunings and the associated error calculated with test sets.
}
\description{
RCV
}
