% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictorRFEAT.R
\name{RF_predictor}
\alias{RF_predictor}
\title{Random Forest + Efficiency Analysis Trees Predictor}
\usage{
RF_predictor(forest, xn)
}
\arguments{
\item{forest}{List containing individual Efficiency Analysis Trees.}

\item{xn}{Row indexes in data.}
}
\value{
Vector of predictions.
}
\description{
This function predicts the expected value based on a set of inputs.
}
