% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEAT.R
\name{RFEAT}
\alias{RFEAT}
\title{Random Forest EAT}
\usage{
RFEAT(data, x, y, numStop = 5, m = 50, s_mtry = "Breiman", na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{numStop}{Integer. Minimun number of observations in a node for a split to be attempted.}

\item{m}{Integer. Number of trees to be build.}

\item{s_mtry}{Select number of inputs in each split. It can be an integer or any of the following options:
\itemize{
\item{\code{"Breiman"}}: \code{in / 3}
\item{\code{"DEA1"}}: \code{(t.obs / 2) - out}  
\item{\code{"DEA2"}}: \code{(t.obs / 3) - out}
\item{\code{"DEA3"}}: \code{t.obs - 2 * out}
\item{\code{"DEA4"}}: \code{min(t.obs / out, (t.obs / 3) - out)}
}}

\item{na.rm}{Logical. If \code{TRUE}, NA rows are omitted.}
}
\value{
A RFEAT object containing:
\itemize{
  \item{\code{data} \itemize{
                      \item{\code{df}}: data frame containing the variables in the model.
                      \item{\code{x}}: input indexes in data.
                      \item{\code{y}}: output indexes in data.
                      \item{\code{input_names}}: input variable names.
                      \item{\code{output_names}}: output variable names.
                      \item{\code{row_names}}: rownames in data.}
       }
  \item{\code{control} \itemize{
                        \item{\code{numStop}}: numStop hyperparameter value.
                        \item{\code{m}}: m hyperparameter value.
                        \item{\code{s_mtry}}: s_mtry hyperparameter value.
                        \item{\code{na.rm}}: na.rm hyperparameter value.}
       }
  \item{\code{forest}: list structure containing the individual EAT.}
  \item{\code{error}: Out-of-Bag error at the forest.}   
  \item{\code{OOB}: list containing Out-of-Bag samples for each tree.}
}
}
\description{
This function build \code{m} individual Efficiency Analysis Trees in a forest structure.
}
\examples{

simulated <- eat:::X2Y2.sim(N = 50, border = 0.1)

RFmodel <- RFEAT(data = simulated, x = c(1,2), y = c(3, 4), numStop = 5,
                  m = 50, s_mtry = "Breiman", na.rm = TRUE)

}
