% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEAT.R
\name{RFEAT}
\alias{RFEAT}
\title{Random Forest EAT}
\usage{
RFEAT(data, x, y, numStop = 5, m = 50, s_mtry = "Breiman", na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{numStop}{Integer. Minimun number of observations in a node for a split to be attempted.}

\item{m}{Integer. Number of trees to be build.}

\item{s_mtry}{Select number of inputs in each split. It can be an integer or any of the following options:
\itemize{
\item{\code{"Breiman"}}: \code{in / 3}
\item{\code{"DEA1"}}: \code{(t.obs / 2) - out}  
\item{\code{"DEA2"}}: \code{(t.obs / 3) - out}
\item{\code{"DEA3"}}: \code{t.obs - 2 * out}
\item{\code{"DEA4"}}: \code{min(t.obs / out, (t.obs / 3) - out)}
}}

\item{na.rm}{Logical. If \code{TRUE}, NA rows are omitted.}
}
\value{
A RFEAT object.
}
\description{
This function build \code{m} individual Efficiency Analysis Trees in a forest structure.
}
\examples{

simulated <- eat:::X2Y2.sim(N = 50, border = 0.1)

RFmodel <- RFEAT(data = simulated, x = c(1,2), y = c(3, 4), numStop = 5,
                  m = 50, s_mtry = "Breiman", na.rm = TRUE)

}
