% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfEAT.R
\name{RFEAT}
\alias{RFEAT}
\title{Random Forest EAT}
\usage{
RFEAT(data, x, y, numStop = 5, m, s_mtry = "Breiman", na.rm = TRUE)
}
\arguments{
\item{data}{Dataframe or matrix containing the variables in the model.}

\item{x}{Vector. Column input indexes in data.}

\item{y}{Vector. Column output indexes in data.}

\item{numStop}{Integer. Minimun number of observations in a node for a split to be attempted.}

\item{m}{Integer. Number of trees to be build.}

\item{s_mtry}{Select number of inputs. It could be: \code{"Breiman"}, \code{"DEA1"}, \code{"DEA2"}, \code{"DEA3"} or \code{"DEA4"}.}

\item{na.rm}{Logical. If \code{TRUE}, NA rows are omitted.}
}
\value{
List of m trees in forest and the error that will be used in the ranking of the importance of the variables.
}
\description{
This function build "m" EAT trees in a forest structure.
}
