% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitRFEAT.R
\name{select_mtry}
\alias{select_mtry}
\title{Select possible inputs in split.}
\usage{
select_mtry(s_mtry, t, nX, nY)
}
\arguments{
\item{s_mtry}{Select number of inputs. It could be: \code{"BRM"}, \code{"DEA1"}, \code{"DEA2"}, \code{"DEA3"} or \code{"DEA4"} or any integer.}

\item{t}{Node which is being splitted.}

\item{nX}{Number of inputs in data.}

\item{nY}{Number of outputa in data.}
}
\value{
Number of inputs selected according with the rule specified.
}
\description{
This function selects the number of inputs for a split in Random Forest.
}
