% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAT.R
\name{size}
\alias{size}
\title{Efficiency Analysis Trees Size}
\usage{
size(object)
}
\arguments{
\item{object}{An EAT object.}
}
\value{
Print with the number of leaf nodes at the tree.
}
\description{
This function returns the number of leaf nodes at the tree.
}
\examples{

simulated <- Y1.sim(N = 50, nX = 3)
model <- EAT(data = simulated, x = c(1, 2, 3), y = 4, numStop = 10, fold = 5)
size(model)

}
