% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitRFEAT.R
\name{split_forest}
\alias{split_forest}
\title{Split Node in Random Forest EAT}
\usage{
split_forest(data, tree, leaves, t, x, y, numStop, arrayK)
}
\arguments{
\item{data}{Data to be used.}

\item{tree}{List structure with the tree nodes.}

\item{leaves}{List with leaf nodes or pending expansion nodes.}

\item{t}{Node which is being split.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{numStop}{Minimum number of observations on a node to be split.}

\item{arrayK}{Column input indexes in data selected by s_mtry.}
}
\value{
Leaves and tree lists updated with the new child nodes.
}
\description{
This function gets the variable and split value to be used in estimEAT, selects the best split, node indexes and leaf list.
}
