% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/training.R
\name{training}
\alias{training}
\title{Estimate Agent Model}
\usage{
training(trainData, features, Formula, k, sampling = FALSE,
  package = c("caretglm", "caretglmnet", "glm", "caretnnet", "caretdnn"),
  parallel = FALSE)
}
\arguments{
\item{trainData}{data.frame dataset with individual decisions.}

\item{features}{list}

\item{Formula}{list}

\item{k}{numeric vector length one}

\item{sampling}{= FALSE}

\item{package}{optional charac vector length one: "caretglm", "caretglmnet", "glm", "caretnnet", "caretdnn"}

\item{parallel}{= FALSE}
}
\value{
Returns a list.
}
\description{
\code{training} uses caret to estimate a model
}

