% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesForRCV.R
\name{treesForRCV}
\alias{treesForRCV}
\title{Trees for RCV}
\usage{
treesForRCV(notLv, x, y, fold, numStop)
}
\arguments{
\item{notLv}{Train set.}

\item{x}{Column input indexes in data.}

\item{y}{Column output indexes in data.}

\item{fold}{Parts in which the original set is divided to perform Cross-Validation.}

\item{numStop}{Minimum number of observations in a node to be split.}
}
\value{
List with each possible pruning for the deep tree generated with train set and its associated alpha values.
}
\description{
This function generates a deep EAT and all pruning for each train set.
}
