% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_program.R
\name{what_outcome}
\alias{what_outcome}
\title{Determine the class of the outcome}
\usage{
what_outcome(y)
}
\arguments{
\item{y}{outcome vector
  
Trying to determine if this is a > 2 class problem.}
}
\value{
Returns a list with two elements: \code{outcome_class} and
  \code{levels}.
}
\description{
Determine the class of the outcome
}

