% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{Y1.sim}
\alias{Y1.sim}
\title{Single Output Data Generation}
\usage{
Y1.sim(N, nX)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{nX}{Integer. Number of inputs. \code{1}, \code{3}, \code{6}, \code{9}, \code{12} and \code{15} are acceptable.}
}
\value{
Data frame with the values of a scenario with nX inputs and 1 output.
}
\description{
This function is used to simulate the data in a single output scenario.
}
